Moved some UI code into its own file
