+++ .vscode/launch.json
@@ -9,7 +9,7 @@
             "args": [
             ],
             "stopAtEntry": false,
-            "cwd": "${workspaceFolder}/build/Debug/inst/bin",
+            "cwd": "${workspaceFolder}/build_vscode/inst/bin",
             "environment": [],
             "externalConsole": false,
             "MIMode": "gdb",
@@ -19,11 +19,11 @@
             "name": "Unit Tests",
             "type": "cppdbg",
             "request": "launch",
-            "program": "${workspaceFolder}/build/Debug/inst/bin/remc2-unit-test",
+            "program": "${workspaceFolder}/build_vscode/inst/bin/remc2-unit-test",
             "args": [
             ],
             "stopAtEntry": false,
-            "cwd": "${workspaceFolder}/build/Debug/inst/bin",
+            "cwd": "${workspaceFolder}/build_vscode/inst/bin",
             "environment": [],
             "externalConsole": false,
             "MIMode": "gdb",
@@ -33,13 +33,13 @@
             "name": "Renderer Test",
             "type": "cppdbg",
             "request": "launch",
-            "program": "${workspaceFolder}/build/Debug/inst/bin/remc2",
+            "program": "${workspaceFolder}/build_vscode/inst/bin/remc2",
             "args": [
                 "--set_level", "0",
                 "--test_renderers"
             ],
             "stopAtEntry": false,
-            "cwd": "${workspaceFolder}/build/Debug/inst/bin",
+            "cwd": "${workspaceFolder}/build_vscode/inst/bin",
             "environment": [],
             "externalConsole": false,
             "MIMode": "gdb",
@@ -49,11 +49,11 @@
             "name": "Regression Tests",
             "type": "cppdbg",
             "request": "launch",
-            "program": "${workspaceFolder}/build/Debug/inst/bin/remc2-regression-test",
+            "program": "${workspaceFolder}/build_vscode/inst/bin/remc2-regression-test",
             "args": [
             ],
             "stopAtEntry": false,
-            "cwd": "${workspaceFolder}/build/Debug/inst/bin",
+            "cwd": "${workspaceFolder}/build_vscode/inst/bin",
             "environment": [],
             "externalConsole": false,
             "MIMode": "gdb",
