+++ .vscode/settings.json
@@ -4,5 +4,6 @@
         "iostream": "cpp",
         "cstdlib": "cpp",
         "filesystem": "cpp"
-    }
+    },
+    "C_Cpp.default.configurationProvider": "ms-vscode.cmake-tools"
 }
