+++ remc2/engine/GameUI.cpp
@@ -1,0 +1,3619 @@
+#include "GameUI.h"
+
+#include <array>
+
+#include "Basic.h"
+#include "Terrain.h"
+#include "GameBitmapIndexes.h"
+#include "LangTextIndexes.h"
+#include "Type_D93C0_Bldgprmbuffer.h"
+
+void ColorizeScreen_2E850(int posX, int posY, int width, int height, uint8_t color);
+
+char x_BYTE_D419D_fonttype = 1; // weak
+char* SpellLevelText_DB06C[5] = { (char*)"I",(char*)"II",(char*)"III",(char*)"IV",(char*)"V" };//x_DWORD_DAF50ar[0x11c]//2ac06c
+std::array<uint8_t, 25> IndexLevelText_DB4EE = //x_DWORD_DAF50ar[0x59e] //index for levels endings?
+{
+	0x30, 0x36, 0x3C, 0x42, 0x45,
+	0x48, 0x4D, 0x4F, 0x56, 0x5C,
+	0x61, 0x66, 0x69, 0x6E, 0x73,
+	0x76, 0x7C, 0x7E, 0x83, 0x85,
+	0x88, 0x8C, 0x8F, 0x97, 0x9C
+};
+std::array<uint8_t, 25> LevelEndText_DB507 = ////index for levels endings?
+{ 
+	0x35, 0x3B, 0x41, 0x44, 0x47, 
+	0x4C, 0x4E, 0x55, 0x5B, 0x60, 
+	0x65, 0x68, 0x6D, 0x72, 0x75, 
+	0x7B, 0x7D, 0x82, 0x84, 0x87, 
+	0x8B, 0x8E, 0x96, 0x9C, 0x9E 
+};
+
+std::array<uint8_t, 24> x_BYTE_DB520 = 
+{ //index for levels endings?
+	0x30, 0x33, 0x3F, 0x3D, 0x92, 0x71,
+	0x4E, 0x04, 0x01, 0x00, 0x02, 0x00,
+	0x08, 0x00, 0x10, 0x00, 0x20, 0x00,
+	0x80, 0x00, 0x00, 0x01, 0x00, 0x04 
+};
+
+void(*ptrDrawBitmap_F01E8)(int16_t, int16_t, bitmap_pos_struct_t, uint8_t scale); // eax
+
+int16_t x_WORD_F4960[1664 * 8]; // Was 1664. Seems to be used in rendering the map 2x indexes ar a time. fix it -  weak
+
+
+char* x_InterlockedExchange(char**  Target, char*  Value)
+{
+	char* temp = *Target;
+	*Target = Value;
+	return temp;
+};
+
+//----- (0002BD10) --------------------------------------------------------
+void sub_2BD10_draw_line(int16_t a1, int16_t a2, int16_t a3, int16_t a4, uint8_t a5)//20cd10
+{
+	std::function<void(uint16_t, uint16_t, uint16_t, uint16_t, char)> func_draw;
+
+	//void* v5; // ebx
+
+	//int result; // eax
+	uint8_t* temp_screen_buffer; // ST14_4
+
+	if (x_WORD_180660_VGA_type_resolution & 1)
+		func_draw = &sub_BD524;
+	else
+		func_draw = &sub_BD542;
+	func_draw(a1, a2, a3, a4, a5);
+
+	//result = (int)x_D41A0_BYTEARRAY_0;
+	if (D41A0_0.m_GameSettings.m_Display.m_uiScreenSize == 1)
+	{
+		temp_screen_buffer = pdwScreenBuffer_351628;
+		pdwScreenBuffer_351628 = (uint8_t*)x_DWORD_E9C3C;
+		func_draw(a1, a2, a3, a4, a5);
+		//result = v7;
+		pdwScreenBuffer_351628 = temp_screen_buffer;
+	}
+	//return result;
+}
+// D41A0: using guessed type int x_D41A0_BYTEARRAY_0;
+// E9C3C: using guessed type int x_DWORD_E9C3C;
+// 180628: using guessed type int pdwScreenBuffer_351628;
+// 180660: using guessed type __int16 x_WORD_180660_VGA_type_resolution;
+
+//----- (0002D190) --------------------------------------------------------
+void DrawBar_2D190(int16_t posStartX, int16_t posStartY, int16_t maxPosEndX, int16_t posEndY, int16_t posEndX, uint8 colorIdx)//20e190
+{
+	if (posEndX >= maxPosEndX)
+		posEndX = maxPosEndX;
+	if (posEndX >= 2)
+		DrawLine_2BC80(posStartX, posStartY, posEndX, posEndY, colorIdx);
+}
+
+//----- (0002D710) --------------------------------------------------------
+void DrawTopStatusBar_2D710(type_event_0x6E8E* a1x, uint8_t scale)//20e710
+{
+    int16_t posX; // bx
+	type_event_0x6E8E* v2x; // ecx
+	int v3; // ecx
+	//int v4; // edx
+	int v5; // eax
+	//__int64 v6; // rtt
+	type_event_0x6E8E* v9x; // eax
+	uint8_t v16; // [esp-4h] [ebp-4Ch]
+	int v17; // [esp+0h] [ebp-48h]
+	int v18; // [esp+8h] [ebp-40h]
+	//type_str_0x6E8E* v19x; // [esp+Ch] [ebp-3Ch]
+	//int v20; // [esp+10h] [ebp-38h]
+	int16_t v21; // [esp+1Ch] [ebp-2Ch]
+	int16_t barStartXPos; // [esp+24h] [ebp-24h]
+	type_event_0x6E8E* v23x; // [esp+28h] [ebp-20h]
+	int16_t v24; // [esp+2Ch] [ebp-1Ch]
+	int16_t v25; // [esp+30h] [ebp-18h]
+	uint8_t v26; // [esp+34h] [ebp-14h]
+	uint8_t manaMaxColourIdx; // [esp+3Ch] [ebp-Ch]
+	uint8_t manaColourIdx; // [esp+40h] [ebp-8h]
+
+	manaColourIdx = x_BYTE_E88E0x[3 * GetTrueWizardNumber_61790(a1x->dword_0xA4_164x->word_0x38_56)];
+	manaMaxColourIdx = x_BYTE_E88E0x[1 + 3 * GetTrueWizardNumber_61790(a1x->dword_0xA4_164x->word_0x38_56)];
+	v26 = x_BYTE_E88E0x[3 * GetTrueWizardNumber_61790(a1x->dword_0xA4_164x->word_0x38_56) + x_D41A0_BYTEARRAY_4_struct.byteindex_121[2]];
+	//ptrDrawBitmap_F01E8(2, 2, **filearray_2aa18c[6] + 240);
+
+	//Panel around MiniMap
+	ptrDrawBitmap_F01E8((2 * scale), (2 * scale), (*filearray_2aa18c[filearrayindex_MSPRD00DATTAB].posistruct)[MINI_MAP_PANEL], scale);
+
+	//v1 = *(unsigned __int8 *)(*filearray_2aa18c[6] + 244) + 2;
+	posX = ((*filearray_2aa18c[filearrayindex_MSPRD00DATTAB].posistruct)[MINI_MAP_PANEL].width_4 * scale) + (2 * scale);
+	//adress 20e7db
+	v18 = x_D41A0_BYTEARRAY_4_struct.str_index_242ar.dword_4;
+	//v20 = a1x->dword_0xA4_164;
+	v2x = x_DWORD_EA3E4[a1x->dword_0xA4_164x->word_0x3A_58];
+	v23x = v2x;
+	if (v2x > x_DWORD_EA3E4[0] && v2x->dword_0x10_16)
+	{
+		if (a1x->dword_0xA4_164x->byte_0x195_405 > 0 && x_D41A0_BYTEARRAY_4_struct.byteindex_121[1])
+		{
+			//ptrDrawBitmap_F01E8(v1, 2, **filearray_2aa18c[6] + 330);
+			//Draw Castle Health Panel Hit
+			ptrDrawBitmap_F01E8(posX, (2 * scale), (*filearray_2aa18c[filearrayindex_MSPRD00DATTAB].posistruct)[HEALTH_PANEL_HIT], scale);
+			a1x->dword_0xA4_164x->byte_0x195_405--;
+		}
+		else
+		{
+			//Draw Castle Health Panel
+			ptrDrawBitmap_F01E8(posX, (2 * scale), (*filearray_2aa18c[filearrayindex_MSPRD00DATTAB].posistruct)[HEALTH_PANEL], scale);
+		}
+		//sub_2BB40_draw_bitmap(v1 + 2, 2, (uint8_t**)(6 * (v23[4] + 43) + **filearray_2aa18c[6]));
+		
+		//Castle Icon
+		sub_2BB40_draw_bitmap(posX + (2 * scale), (2 * scale), (*filearray_2aa18c[filearrayindex_MSPRD00DATTAB].posistruct)[v23x->dword_0x10_16 + 43], scale);
+		//Castle Health And Mana Icon
+		sub_2BB40_draw_bitmap(posX + (38 * scale), (2 * scale), (*filearray_2aa18c[filearrayindex_MSPRD00DATTAB].posistruct)[HEALTH_MANA_ICONS], scale);
+		barStartXPos = posX + (58 * scale);
+		DrawBar_2D190(posX + (58 * scale), 10 * scale, 64 * scale, 10 * scale, (v23x->life_0x8 << 6) / v23x->maxLife_0x4, 0x7Bu);
+		v17 = (v23x->mana_0x90_144 + a1x->dword_0xA4_164x->dword_0x13C_316) << 6;
+		if (v23x->mana_0x90_144 + a1x->dword_0xA4_164x->dword_0x13C_316 == v23x->maxMana_0x8C_140)
+		{
+			if (x_D41A0_BYTEARRAY_4_struct.byteindex_121[2] == 1)
+			{
+				v3 = v18;
+				//v4 = v23x->dword_0x8C_140 << 6;
+				v16 = manaMaxColourIdx;
+				v5 = v23x->maxMana_0x8C_140 << 6;
+			}
+			else
+			{
+				v16 = manaColourIdx;
+				v5 = (v23x->mana_0x90_144 + a1x->dword_0xA4_164x->dword_0x13C_316) << 6;
+				//v4 = (v23x->dword_0x90_144 + a1x->dword_0xA4_164x->dword_0x13C_316) << 6;
+				v3 = v18;
+			}
+		}
+		else
+		{
+			DrawBar_2D190(barStartXPos, 28 * scale, 64 * scale, 10 * scale, ((v23x->maxMana_0x8C_140 << 6) / v18) * scale, manaMaxColourIdx);
+			v16 = manaColourIdx;
+			v5 = v17;
+			//v4 = v17;
+			v3 = v18;
+		}
+		//LODWORD(v6) = v5;
+		//HIDWORD(v6) = v4 >> 31;
+		DrawBar_2D190(barStartXPos, 28 * scale, 64 * scale, 10 * scale, /*v6*/(v5 / v3) * scale, v16);
+
+		//Draw Mana Goal Lines
+		if (!(x_D41A0_BYTEARRAY_4_struct.setting_byte1_22 & 0x10) && D41A0_0.terrain_2FECE.word_0x2FED5)
+		{
+			DrawLine_2BC80(barStartXPos + (D41A0_0.terrain_2FECE.word_0x2FED5 << 6) / 100, 26 * scale, 2 * scale, 2 * scale, v26);
+			DrawLine_2BC80(barStartXPos + (D41A0_0.terrain_2FECE.word_0x2FED5 << 6) / 100, 38 * scale, 2 * scale, 2 * scale, v26);
+		}
+	}
+	else
+	{
+		//Panel Next to MiniMap
+		ptrDrawBitmap_F01E8(posX, (2 * scale), (*filearray_2aa18c[filearrayindex_MSPRD00DATTAB].posistruct)[PANEL], scale);
+	}
+	//v8 = *(unsigned __int8 *)(**filearray_2aa18c[6] + 250) + v1;
+	//adress 20eaa0
+	posX = (*filearray_2aa18c[filearrayindex_MSPRD00DATTAB].posistruct)[HEALTH_PANEL].width_4 * scale + posX;
+	if (x_DWORD_EA3E4[0] >= v23x)
+	{
+		//20eca6
+		//VGA_Debug_Blit(640, 480, pdwScreenBuffer_351628);
+	    //2nd Panel 
+		ptrDrawBitmap_F01E8(posX, (2 * scale), (*filearray_2aa18c[filearrayindex_MSPRD00DATTAB].posistruct)[PANEL], scale);
+		//VGA_Debug_Blit(640, 480, pdwScreenBuffer_351628);
+	}
+	else
+	{
+		if (a1x->dword_0xA4_164x->byte_0x197_407 > 0 && x_D41A0_BYTEARRAY_4_struct.byteindex_121[1])
+		{
+			//Draw Balloon Health Panel Hit
+			ptrDrawBitmap_F01E8(posX, (2 * scale), (*filearray_2aa18c[filearrayindex_MSPRD00DATTAB].posistruct)[HEALTH_PANEL_HIT], scale);
+			a1x->dword_0xA4_164x->byte_0x197_407--;
+		}
+		else
+		{
+			//Draw Balloon Health Panel Hit
+			ptrDrawBitmap_F01E8(posX, (2 * scale), (*filearray_2aa18c[filearrayindex_MSPRD00DATTAB].posistruct)[HEALTH_PANEL], scale);
+		}
+		switch (v23x->dword_0x10_16)
+		{
+		case 1:
+		case 2:
+		case 3:
+			v25 = 1;
+			break;
+		case 4:
+		case 5:
+			v25 = 2;
+			break;
+		case 6:
+		case 7:
+			v25 = 3;
+			break;
+		default:
+			v25 = 0;
+			break;
+		}
+		//Draw Balloon Icon
+		sub_2BB40_draw_bitmap(posX + (2 * scale), (2 * scale), (*filearray_2aa18c[filearrayindex_MSPRD00DATTAB].posistruct)[v25 + 50], scale);
+		v24 = 0;
+		//Draw Balloon Health Icon
+		sub_2BB40_draw_bitmap(posX + (38 * scale), (2 * scale), (*filearray_2aa18c[filearrayindex_MSPRD00DATTAB].posistruct)[HEALTH_MANA_ICONS], scale);
+		if (v25 > 0)
+		{
+			do
+			{
+				v21 = posX + (58 * scale);
+				v9x = x_DWORD_EA3E4[a1x->dword_0xA4_164x->array_0x3C_60[v24]];
+				//v19x = v9x;
+				if (v9x > x_DWORD_EA3E4[0])
+				{
+					if (v9x->life_0x8 >= 0)
+						DrawBar_2D190(v21, 2 * v24 + 12 * scale, 64 * scale, 2 * scale, ((v9x->life_0x8 << 6) / v9x->maxLife_0x4) * scale, 0x7Bu);
+					DrawBar_2D190(v21, 2 * v24 + 30 * scale, 64 * scale, 2 * scale, ((v9x->mana_0x90_144 << 6) / v9x->maxMana_0x8C_140) * scale, manaColourIdx);
+				}
+				v24++;
+			} while (v24 < v25);
+		}
+	}
+	posX = (*filearray_2aa18c[filearrayindex_MSPRD00DATTAB].posistruct)[HEALTH_PANEL].width_4 * scale + posX;
+
+	if (a1x->dword_0xA4_164x->byte_0x196_406 > 0 && x_D41A0_BYTEARRAY_4_struct.byteindex_121[1])
+	{
+		//Player Health Panel on Hit
+		ptrDrawBitmap_F01E8(posX, 2 * scale, (*filearray_2aa18c[filearrayindex_MSPRD00DATTAB].posistruct)[HEALTH_PANEL_HIT], scale);
+		a1x->dword_0xA4_164x->byte_0x196_406--;
+	}
+	else
+	{//adress 20ed35
+		//Player Health Panel
+		ptrDrawBitmap_F01E8(posX, 2 * scale, (*filearray_2aa18c[filearrayindex_MSPRD00DATTAB].posistruct)[HEALTH_PANEL], scale);
+	}
+
+	if ((a1x->struct_byte_0xc_12_15.byte[0] & 0x20 || a1x->dword_0xA4_164x->word_0x159_345) && x_D41A0_BYTEARRAY_4_struct.byteindex_121[2])
+	{
+		//Draw Player Health and Mana flickering
+		GameBitmap::DrawTransparentBitmap_2DE80(posX + (2 * scale), (2 * scale), (*filearray_2aa18c[filearrayindex_MSPRD00DATTAB].posistruct)[PLAYER_ICON], scale);
+		GameBitmap::DrawTransparentBitmap_2DE80(posX + (38 * scale), (2 * scale), (*filearray_2aa18c[filearrayindex_MSPRD00DATTAB].posistruct)[HEALTH_MANA_ICONS], scale);
+	}
+	else
+	{
+		//Draw Player Health and Mana
+		sub_2BB40_draw_bitmap(posX + (2 * scale), (2 * scale), (*filearray_2aa18c[filearrayindex_MSPRD00DATTAB].posistruct)[PLAYER_ICON], scale);
+		sub_2BB40_draw_bitmap(posX + (38 * scale), (2 * scale), (*filearray_2aa18c[filearrayindex_MSPRD00DATTAB].posistruct)[HEALTH_MANA_ICONS], scale);
+	}
+	//VGA_Debug_Blit(640, 480, pdwScreenBuffer_351628);
+	posX = posX + (58 * scale);
+	DrawBar_2D190(posX, 10 * scale, 64 * scale, 10 * scale, ((a1x->life_0x8 << 6) / a1x->maxLife_0x4) * scale, 0x7Bu);
+	//adress 20ee1a
+	//1b8 1c 40 0a 03 aa
+	//0xfa00/4c79=3
+	//v18=0x4c79
+	DrawBar_2D190(posX, 28 * scale, 64 * scale, 10 * scale, ((a1x->maxMana_0x8C_140 << 6) / v18) * scale, manaMaxColourIdx);
+	DrawBar_2D190(posX, 28 * scale, 64 * scale, 10 * scale, ((a1x->mana_0x90_144 << 6) / v18) * scale, manaColourIdx);
+	//VGA_Debug_Blit(640, 480, pdwScreenBuffer_351628);
+}
+
+// D41A0: using guessed type int x_D41A0_BYTEARRAY_0;
+// D41A4: using guessed type int x_DWORD_D41A4;
+// EA3DC: using guessed type int **filearray_2aa18c[6];
+// EA3E4: using guessed type int x_DWORD_EA3E4[];
+// F01E8: using guessed type int (*ptrDrawBitmap_F01E8)(x_DWORD, x_DWORD, x_DWORD);
+// 180628: using guessed type int pdwScreenBuffer_351628;
+// 18062C: using guessed type int screenWidth_18062C;
+// 180660: using guessed type __int16 x_WORD_180660_VGA_type_resolution;
+
+//----- (0002E260) --------------------------------------------------------
+void DrawSpellIcon_2E260(int16_t posX, int16_t posY, type_event_0x6E8E* a3x, char a4, uint8_t scale)//20f260
+{
+	type_event_0x6E8E* v4x; // eax
+	signed int v5; // eax
+	//__int16 v6; // t1
+	__int16 v7; // cx
+	bitmap_pos_struct_t v8; // eax
+	bitmap_pos_struct_t v9; // eax
+	int v10; // edi
+	unsigned __int16 v11; // di
+	unsigned __int8 v12; // al
+	signed int i; // [esp+0h] [ebp-14h]
+	signed int j; // [esp+4h] [ebp-10h]
+	type_event_0x6E8E* v15x; // [esp+8h] [ebp-Ch]
+	unsigned __int8 v16; // [esp+Ch] [ebp-8h]
+	unsigned __int8 v17; // [esp+10h] [ebp-4h]
+
+	if (a3x > x_DWORD_EA3E4[0])
+	{
+		GetFont_6FC50(x_BYTE_D419D_fonttype);
+		v4x = x_DWORD_EA3E4[a3x->parentId_0x28_40];
+		v15x = v4x;
+		if (v4x > x_DWORD_EA3E4[0])
+		{
+			v17 = x_BYTE_E88E0x[3 * GetTrueWizardNumber_61790(v4x->dword_0xA4_164x->word_0x38_56)];
+			v5 = GetTrueWizardNumber_61790(v15x->dword_0xA4_164x->word_0x38_56);
+			//v6 = a3x->word_0x2E_46;
+			v16 = x_BYTE_E88E0x[1 + 3 * v5];
+			//if (!(SPELLS_BEGIN_BUFFER_DA818[0x1+80 * a3x->byte_0x40_64] & 4)
+			if (!(SPELLS_BEGIN_BUFFER_str[a3x->model_0x40_64].byte_1 & 4)
+				|| (v7 = a3x->word_0x2E_46, v7 <= 0)
+				|| v7 >= 32
+				|| !x_D41A0_BYTEARRAY_4_struct.byteindex_121[1])
+			{
+				if (a4)//adress 20f334
+				{
+					if (a3x->word_0x2E_46)
+						//v8 = (uint8_t**)(**filearray_2aa18c[6] + 12);
+						v8 = (*filearray_2aa18c[filearrayindex_MSPRD00DATTAB].posistruct)[2];
+					else
+						//v8 = (uint8_t**)(**filearray_2aa18c[6] + 6);
+						v8 = (*filearray_2aa18c[filearrayindex_MSPRD00DATTAB].posistruct)[1];
+					sub_2BB40_draw_bitmap(posX, posY, v8, scale);
+				}
+				else
+				{
+					if (a3x->word_0x2E_46)
+						//v9 = **filearray_2aa18c[6] + 12;
+						v9 = (*filearray_2aa18c[filearrayindex_MSPRD00DATTAB].posistruct)[2];
+					else
+						//v9 = **filearray_2aa18c[6] + 6;
+						v9 = (*filearray_2aa18c[filearrayindex_MSPRD00DATTAB].posistruct)[1];
+					ptrDrawBitmap_F01E8(posX, posY, v9, scale);
+				}//20f38d
+				//a6757c+2e8=a67864->a60a57
+				sub_2BB40_draw_bitmap(posX, posY, (*filearray_2aa18c[filearrayindex_MSPRD00DATTAB].posistruct)[(a3x->model_0x40_64 + 123)], scale);
+				//20f3b8
+				//"I" 0x234 a6757c->a5b3b0    0xc
+				//adress 20f37a
+				DrawText_2BC10((char*)SpellLevelText_DB06C[a3x->byte_0x46_70],
+					(*filearray_2aa18c[filearrayindex_MSPRD00DATTAB].posistruct)[1].width_4 * scale + posX - (8 * scale) * strlen((const char*)*(&SpellLevelText_DB06C[a3x->byte_0x46_70])) - (2 * scale),//test and fix
+					posY,
+					(*xadataclrd0dat.colorPalette_var28)[0],
+					scale);//x_BYTE_E8900//20f3bd//2b9900
+				if (a3x->maxMana_0x8C_140)
+				{
+					DrawLine_2BC80(
+						posX + (4 * scale),
+						posY + (36 * scale),
+						(56 * scale) * (v15x->mana_0x90_144 % a3x->maxMana_0x8C_140) / a3x->maxMana_0x8C_140,
+						(4 * scale),
+						v16);
+					v10 = v15x->mana_0x90_144 / a3x->maxMana_0x8C_140;
+					for (i = 0; i < 27 && v10; i++)
+					{
+						for (j = 0; j < 2 && v10; j++)
+						{
+							DrawLine_2BC80(posX + (2 * (i + 2) * scale), posY + (2 * (j + 18) * scale), 2 * scale, 2 * scale, v17);
+							v10--;
+						}
+					}
+				}
+				if (a3x->dword_0x88_136)
+				{
+					v11 = v15x->dword_0xA4_164x->word_0x3A_58;
+					if (!v11 || a3x->dword_0x88_136 > x_DWORD_EA3E4[v11]->mana_0x90_144)
+					{
+						if (D41A0_0.terrain_2FECE.MapType != MapType_t::Day)
+							v12 = 16;
+						else
+							v12 = 48;
+						ColorizeScreen_2E850(posX, posY, (*filearray_2aa18c[filearrayindex_MSPRD00DATTAB].posistruct)[1].width_4 * scale, (*filearray_2aa18c[filearrayindex_MSPRD00DATTAB].posistruct)[1].height_5 * scale, v12);
+					}
+				}
+			}
+		}
+	}
+}
+// D419D: using guessed type char x_BYTE_D419D_fonttype;
+// D41A0: using guessed type int x_D41A0_BYTEARRAY_0;
+// D41A4: using guessed type int x_DWORD_D41A4;
+// DB06C: using guessed type void *off_DB06C;
+// EA3DC: using guessed type int **filearray_2aa18c[6];
+// EA3E4: using guessed type int x_DWORD_EA3E4[];
+// F01E8: using guessed type int (*ptrDrawBitmap_F01E8)(x_DWORD, x_DWORD, x_DWORD);
+
+//----- (00089830) --------------------------------------------------------
+void ComputeTextboxSizes_89830(Type_TextBox_1804B0* textbox, uint8_t scale)//26a830
+{
+	int16_t preMaxTextboxWidth = textbox->lineSrcX_0x1c + (25 * scale);
+	if (textbox->maxPosX_0x16 < textbox->lineSrcX_0x1c + (135 * scale))
+		preMaxTextboxWidth = textbox->lineSrcX_0x1c - (135 * scale);
+
+	int16_t preMaxTextboxHeight = textbox->lineSrcY_0x1e - (94 * scale);
+	if (textbox->lineSrcY_0x1e - (94 * scale) + (28 * scale) < textbox->minPosY_0x18)
+		preMaxTextboxHeight = textbox->lineSrcY_0x1e + (66 * scale);
+
+	textbox->maxTextboxWidth_0x0 = preMaxTextboxWidth;
+	textbox->maxTextboxHeight_0x2 = preMaxTextboxHeight;
+}
+
+//----- (000898A0) --------------------------------------------------------
+void ComputeTextboxLine_898A0(Type_TextBox_1804B0* textbox)//26a8a0
+{
+	int preLineX2;
+	int preLineY2;
+
+	int centerToFrameWidth = textbox->framePosX_0x24;
+
+	if (int lineX1 = textbox->lineSrcX_0x1c; lineX1 >= centerToFrameWidth - 30)
+	{
+		if (lineX1 <= textbox->frameWidth_0x28 + centerToFrameWidth + 30)
+			preLineX2 = textbox->maxTextboxWidth_0x0;
+		else
+			preLineX2 = textbox->frameWidth_0x28 + textbox->framePosX_0x24;
+	}
+	else
+	{
+		preLineX2 = textbox->framePosX_0x24 - 2;
+	}
+	textbox->lineDestX_0x20 = preLineX2;
+	int centerToFrameHeight = textbox->framePosY_0x26;
+
+	if (int lineY1 = textbox->lineSrcY_0x1e; lineY1 >= centerToFrameHeight - 30)
+	{
+		if (lineY1 > textbox->frameHeight_0x2a + centerToFrameHeight + 30)
+		{
+			textbox->lineDestY_0x22 = textbox->frameHeight_0x2a + textbox->framePosY_0x26;
+			return;
+		}
+		preLineY2 = textbox->maxTextboxHeight_0x2;
+	}
+	else
+	{
+		preLineY2 = textbox->framePosY_0x26 - 2;
+	}
+	textbox->lineDestY_0x22 = preLineY2;
+}
+
+//----- (00089920) --------------------------------------------------------
+void ComputeTextboxSizesFromTextLines_89920(Type_TextBox_1804B0* textbox, int16_t countLines, int16_t* textLines, uint8_t scale)//26a920
+{
+	int maxTextLenght = 0;
+	int textLenght;
+	for (int indexLine = 0; indexLine < countLines; indexLine++)
+	{
+		textLenght = strlen((const char*)x_DWORD_E9C4C_langindexbuffer[textLines[indexLine]]) + 1;
+		// FireBall | Rapid Fire | Fire Storm		
+		if ((textLenght - 1) > maxTextLenght)
+			maxTextLenght = textLenght - 1;
+	}
+	textbox->textBoxWidth_0x4 = (textbox->charWidth_0x10 * scale) * maxTextLenght;
+	textbox->textboxHeight_0x6 = (textbox->charHeight_0x12 * scale) * countLines;
+}
+
+//----- (00089A30) --------------------------------------------------------
+void DrawTextboxLine_89A30(const Type_TextBox_1804B0* textbox)//26aa30
+{
+	int widthFix = ((x_WORD_180660_VGA_type_resolution & 1) != 0) + 1;
+	if (textbox->lineDestX_0x20 != textbox->maxTextboxWidth_0x0 || textbox->lineDestY_0x22 != textbox->maxTextboxHeight_0x2)
+	{
+		sub_2BD10_draw_line(textbox->lineSrcX_0x1c, textbox->lineSrcY_0x1e, textbox->lineDestX_0x20, textbox->lineDestY_0x22, *xadataclrd0dat.colorPalette_var28[0]);
+		sub_2BD10_draw_line(textbox->lineSrcX_0x1c + widthFix, textbox->lineSrcY_0x1e, widthFix + textbox->lineDestX_0x20, textbox->lineDestY_0x22, (*xadataclrd0dat.colorPalette_var28)[3840]);
+	}
+}
+
+//----- (00089AC0) --------------------------------------------------------
+void GetHintText_89AC0(char* buffer, int helpIndex)//26aac0
+{
+	switch (helpIndex) 
+	{
+	case 92:
+		sprintf(buffer, "[F1] %s", x_DWORD_E9C4C_langindexbuffer[14]);//Toggle help on / off
+		break;
+	case 96:
+		sprintf(buffer, (const char*)x_DWORD_E9C4C_langindexbuffer[NUMBER_OF_TIMES_YOU_HAVE_KILLED], std::string(WizardsNames_D93A0[GetTrueWizardNumber_61790(str_unk_1804B0ar.byte_0xa8)]).c_str() );//Number of times you have killed %s
+		break;
+	default:
+		buffer[0] = 0;
+		break;
+	}
+}
+
+//----- (0002E850) --------------------------------------------------------
+void ColorizeScreen_2E850(int posX, int posY, int width, int height, uint8_t color)//20f850
+{
+	int locPosX = posX;
+	int locPosY = posY;
+	int locWidth = width;
+	int locHeight = height;
+
+	if (x_WORD_180660_VGA_type_resolution == 1)
+	{
+		locPosX = posX / 2;
+		locPosY = posY / 2;
+		locWidth = width / 2;
+		locHeight = height / 2;
+	}
+	for (int indexY = locPosY; indexY < locPosY + locHeight; indexY++)
+	{
+		for (int indexX = locPosX; indexX < locWidth + locPosX; indexX++)
+		{
+			pdwScreenBuffer_351628[screenWidth_18062C * indexY + indexX] = x_BYTE_F6EE0_tablesx[256 * color
+				+ pdwScreenBuffer_351628[screenWidth_18062C * indexY + indexX]];
+		}
+	}
+}
+
+//----- (00030630) --------------------------------------------------------
+void DrawCurrentObjectiveTextbox_30630(uint8_t scale)//211630
+{
+	char v0; // dl
+	int result; // eax
+	unsigned __int16 v2; // cx
+	char* text; // ebx
+	Type_TextBox_1804B0 textbox;
+
+	v0 = x_D41A0_BYTEARRAY_4_struct.setting_byte1_22;
+	result = 0;
+	if (!(v0 & 0x10) && !(v0 & 0x20))
+	{
+		if (x_D41A0_BYTEARRAY_4_struct.setting_byte2_23 & 2)
+		{
+			if (D41A0_0.struct_0x3659C[D41A0_0.LevelIndex_0xc].substr_3659C.IsLevelEnd_0)
+				result = x_BYTE_DB520[6];//Toggle help on / off
+			else
+				result = x_BYTE_DB520[D41A0_0.struct_0x3659C[D41A0_0.LevelIndex_0xc].substr_3659C.ObjectiveText_1];
+		}
+		else if (D41A0_0.byte_0x36E0B & 1)
+		{
+			result = (D41A0_0.array_0x2BDE[D41A0_0.LevelIndex_0xc].byte_0x3E4_2BE4_12226 != 0) + 284;//Invisible
+		}
+		else
+		{
+			v2 = x_D41A0_BYTEARRAY_4_struct.levelnumber_43w;
+			if (v2 >= 0x1Eu && v2 <= 0x22u)
+			{
+				if (D41A0_0.struct_0x3659C[D41A0_0.LevelIndex_0xc].substr_3659C.IsLevelEnd_0)
+					result = LevelEndText_DB507[10];
+				else
+					result = IndexLevelText_DB4EE[0] + 3;
+			}
+			else if (x_D41A0_BYTEARRAY_4_struct.setting_38545 & 0x20
+				&& !D41A0_0.struct_0x3659C[D41A0_0.LevelIndex_0xc].substr_3659C.ObjectiveText_1)
+			{
+				D41A0_0.byte_0x36E04 = 0;
+			}
+			else
+			{
+				if (D41A0_0.struct_0x3659C[D41A0_0.LevelIndex_0xc].substr_3659C.IsLevelEnd_0)
+					result = LevelEndText_DB507[x_D41A0_BYTEARRAY_4_struct.levelnumber_43w];
+				else
+					result = D41A0_0.struct_0x3659C[D41A0_0.LevelIndex_0xc].substr_3659C.ObjectiveText_1
+					+ IndexLevelText_DB4EE[x_D41A0_BYTEARRAY_4_struct.levelnumber_43w];
+			}
+		}
+		if (result)
+		{
+			text = (char*)x_DWORD_E9C4C_langindexbuffer[result];
+			/*"Fly towards my beacon."*/
+
+			textbox.charWidth_0x10 = 8;
+			textbox.maxTextboxHeight2_0xe = 400;
+			textbox.minPosX_0x14 = 40;
+			textbox.maxPosX_0x16 = 600;
+			textbox.charHeight_0x12 = 14;
+			textbox.maxPosY_0x1a = 260;
+
+			if (!DefaultResolutions())
+			{
+				textbox.maxPosX_0x16 = screenWidth_18062C - 40;
+				textbox.maxPosY_0x1a = screenHeight_180624 / 2;
+			}
+
+			textbox.color1_0x30 = (*xadataclrd0dat.colorPalette_var28)[3840];
+			textbox.color2_0x31 = ((*xadataclrd0dat.colorPalette_var28)[0]);
+			textbox.minPosY_0x18 = 80;
+			textbox.maxTextboxWidth_0x0 = 320;
+			textbox.maxTextboxHeight_0x2 = 300;
+			textbox.maxTextboxWidth2_0xc = 320;
+
+			if (!DefaultResolutions())
+			{
+				textbox.maxTextboxWidth_0x0 = screenWidth_18062C / 2;
+				textbox.maxTextboxHeight_0x2 = screenHeight_180624 / 2 - 100;
+				textbox.maxTextboxWidth2_0xc = screenWidth_18062C / 2;
+				textbox.maxTextboxHeight2_0xe = screenHeight_180624 / 2;
+			}
+
+			ComputeTextboxSizesFromTextWords_89420(&textbox, text, scale);
+			ConstrainTextboxSizes_89520(&textbox);
+			ComputeFrameSizes_89980(&textbox, scale);
+			ColorizeScreen_2E790(textbox.framePosX_0x24, textbox.framePosY_0x26, textbox.frameWidth_0x28, textbox.frameHeight_0x2a, textbox.color2_0x31);
+			DrawTextboxText_895D0(&textbox, text, scale);
+			DrawTextboxFrame_89690(&textbox, scale);
+		}
+	}
+}
+
+//----- (000595C0) --------------------------------------------------------
+char sub_595C0(__int16 a1)//23a5c0
+{
+	char result = -1;
+	if (D41A0_0.byte_0x36E0B & 1)
+		return -3;
+	if (!D41A0_0.struct_0x3659C[a1].substr_3659C.IsLevelEnd_0)
+		result = D41A0_0.stages_0x3654C[D41A0_0.struct_0x3659C[a1].substr_3659C.ObjectiveText_1].stages_3654C_byte0;
+	return result;
+}
+
+//----- (00059610) --------------------------------------------------------
+char sub_59610(type_event_0x6E8E** a1y, int16_t a2)//23a610
+{
+	char v5; // dl
+	*a1y = 0;
+	v5 = 0;
+	if (D41A0_0.struct_0x3659C[a2].substr_3659C.IsLevelEnd_0)
+		return v5;
+	if (D41A0_0.stages_0x3654C[D41A0_0.struct_0x3659C[a2].substr_3659C.ObjectiveText_1].stages_3654C_byte0 < 4u)
+	{
+		if (D41A0_0.stages_0x3654C[D41A0_0.struct_0x3659C[a2].substr_3659C.ObjectiveText_1].stages_3654C_byte0 < 1u || D41A0_0.stages_0x3654C[D41A0_0.struct_0x3659C[a2].substr_3659C.ObjectiveText_1].stages_3654C_byte0 > 2u)
+			return v5;
+	}
+	else if (D41A0_0.stages_0x3654C[D41A0_0.struct_0x3659C[a2].substr_3659C.ObjectiveText_1].stages_3654C_byte0 > 4u)
+	{
+		if (D41A0_0.stages_0x3654C[D41A0_0.struct_0x3659C[a2].substr_3659C.ObjectiveText_1].stages_3654C_byte0 == 6 && D41A0_0.stages_0x3654C[D41A0_0.struct_0x3659C[a2].substr_3659C.ObjectiveText_1].str_3654D_byte1 & 1)
+		{
+			v5 = 1;
+			*a1y = x_DWORD_EA3E4[D41A0_0.stages_0x3654C[D41A0_0.struct_0x3659C[a2].substr_3659C.ObjectiveText_1].str_36552_un.dword];
+			return v5;
+		}
+		return v5;
+	}
+	if (D41A0_0.stages_0x3654C[D41A0_0.struct_0x3659C[a2].substr_3659C.ObjectiveText_1].str_3654D_byte1 & 1)
+	{
+		v5 = 1;
+		*a1y = D41A0_0.stages_0x3654C[D41A0_0.struct_0x3659C[a2].substr_3659C.ObjectiveText_1].str_36552_un.ptr0x6E8E;
+	}
+	return v5;
+}
+
+//----- (000596C0) --------------------------------------------------------
+bool sub_596C0(axis_3d* position, int16_t a2)//23a6c0
+{
+	bool result = false;
+	position->x = -1;
+	position->y = -1;
+	position->z = -1;
+	if (!D41A0_0.struct_0x3659C[a2].substr_3659C.IsLevelEnd_0)
+	{
+		if (D41A0_0.stages_0x3654C[D41A0_0.struct_0x3659C[a2].substr_3659C.ObjectiveText_1].stages_3654C_byte0 >= 4u && D41A0_0.stages_0x3654C[D41A0_0.struct_0x3659C[a2].substr_3659C.ObjectiveText_1].stages_3654C_byte0 <= 5u)
+		{
+			position->x = D41A0_0.stages_0x3654C[D41A0_0.struct_0x3659C[a2].substr_3659C.ObjectiveText_1].str_3654E_axis.x;
+			position->y = D41A0_0.stages_0x3654C[D41A0_0.struct_0x3659C[a2].substr_3659C.ObjectiveText_1].str_3654E_axis.y;
+			result = true;
+		}
+	}
+	return result;
+}
+
+//----- (000616D0) --------------------------------------------------------
+signed int sub_616D0(signed int a1)//2426d0
+{
+	signed int result; // eax
+
+	result = 0;
+	if (a1 >= 0 && a1 <= 7)
+	{
+		if (x_D41A0_BYTEARRAY_4_struct.setting_byte1_22 & 0x10)
+		{
+			switch (D41A0_0.array_0x2BDE[a1].dword_0x3E6_2BE4_12228.byte_0x1C0_448)
+			{
+			case 0:
+			LABEL_5:
+				result = 0;
+				break;
+			case 1:
+			LABEL_7:
+				result = 4;
+				break;
+			case 2:
+			LABEL_6:
+				result = 1;
+				break;
+			case 3:
+			LABEL_9:
+				result = 2;
+				break;
+			case 4:
+			LABEL_10:
+				result = 5;
+				break;
+			case 5:
+			case 7:
+			LABEL_13:
+				result = 7;
+				break;
+			case 6:
+			LABEL_8:
+				result = 3;
+				break;
+			default:
+				return result;
+			}
+		}
+		else
+		{
+			switch (a1)
+			{
+			case 0:
+				goto LABEL_5;
+			case 1:
+				goto LABEL_6;
+			case 2:
+				goto LABEL_7;
+			case 3:
+				goto LABEL_8;
+			case 4:
+				goto LABEL_9;
+			case 5:
+				goto LABEL_10;
+			case 6:
+				goto LABEL_13;
+			case 7:
+				result = 6;
+				break;
+			default:
+				return result;
+			}
+		}
+	}
+	return result;
+}
+
+int debugcounter2 = 0;
+//----- (00061A00) --------------------------------------------------------
+void sub_61A00_draw_minimap_entites_b(int16_t x, int16_t y, int16_t posX, int16_t posY, uint16_t width, uint16_t height, int16_t yaw, int16_t scaling)//242a00
+{
+	int v8; // edx
+	int v9; // esi
+	MapType_t v11; // al
+	__int64 v12; // rtt
+	int v13; // ebx
+	int v14; // edx
+	int v15; // ebx
+	__int16 v16; // dx
+	//int v17; // ebx
+	int v18; // eax
+	int v19; // edx
+	int v20; // esi
+	int v21; // edi
+	type_event_0x6E8E* v22x; // edx
+	int v23; // eax
+	int v24; // edx
+	int v25; // ebx
+	int v26; // ST48_4
+	int v27; // edx
+	int i; // eax
+	int v29; // ebx
+	type_event_0x6E8E* jy; // ebx
+	unsigned __int8 v31; // al
+	unsigned __int8 v32; // ah
+	unsigned __int8 v33; // ah
+	unsigned __int8 v34; // ah
+	unsigned __int16 v35; // si
+	type_event_0x6E8E* v36x; // eax
+	char v37; // al
+	char v38; // al
+	type_event_0x6E8E* v39x; // eax
+	unsigned __int16 v40; // si
+	type_event_0x6E8E* v41x; // eax
+	unsigned __int8 v42; // ah
+	char v43; // alv39x
+	unsigned __int8 v44; // ah
+	int v45; // eax
+	unsigned __int8 v46; // al
+	char v47; // al
+	char v48; // dl
+	char v49; // al
+	unsigned __int8 v50; // ah
+	__int16 v51; // ax
+	int v52; // edx
+	int v53; // ecx
+	int v54; // esi
+	int v55; // edi
+	char* v56; // edx
+	int v57; // ecx
+	char* v58; // edx
+	int v59; // eax
+	type_event_0x6E8E* v60x; // ebx
+	int v61; // edx
+	int v62; // eax
+	int v63; // esi
+	int v64; // edi
+	unsigned __int8* v65; // ST14_4
+	unsigned __int8 v66; // al
+	int v67; // esi
+	x_BYTE* v68; // ebx
+	char* index; // eax
+	x_BYTE* v70; // edx
+	int v71; // ecx
+	int v72; // [esp+0h] [ebp-60h]
+	int v73; // [esp+8h] [ebp-58h]
+	type_event_0x6E8E* v74x; // [esp+Ch] [ebp-54h]
+	signed int v75; // [esp+10h] [ebp-50h]
+	int v76; // [esp+14h] [ebp-4Ch]
+	x_BYTE* v77; // [esp+18h] [ebp-48h]
+	signed int v78; // [esp+1Ch] [ebp-44h]
+	//int v79; // [esp+20h] [ebp-40h]
+	type_event_0x6E8E* v80x; // [esp+24h] [ebp-3Ch]
+	x_BYTE* v81; // [esp+28h] [ebp-38h]
+	int v82; // [esp+2Ch] [ebp-34h]
+	int v83; // [esp+30h] [ebp-30h]
+	uint8_t* v84; // [esp+34h] [ebp-2Ch]
+	int v85; // [esp+38h] [ebp-28h]
+	int v86; // [esp+3Ch] [ebp-24h]
+	unsigned __int16 v87; // [esp+40h] [ebp-20h]
+	unsigned __int16 v88; // [esp+44h] [ebp-1Ch]
+	char v89; // [esp+48h] [ebp-18h]
+	uint8_t v90; // [esp+4Ch] [ebp-14h]
+	uint8_t v91; // [esp+50h] [ebp-10h]
+	uint8_t v92; // [esp+54h] [ebp-Ch]
+	uint8_t v93; // [esp+58h] [ebp-8h]
+	char v94; // [esp+5Ch] [ebp-4h]
+
+	if (debugcounter2 >= 8)
+	{
+		debugcounter2 = debugcounter2;
+	}
+	debugcounter2++;
+
+	v8 = x;
+	v9 = y;
+	v11 = D41A0_0.terrain_2FECE.MapType;
+	v83 = 0;
+	if (v11 == MapType_t::Day)
+	{
+		v92 = (*xadataclrd0dat.colorPalette_var28)[0];
+		v93 = (*xadataclrd0dat.colorPalette_var28)[0];
+		v91 = 0xe8;
+		v90 = 0x1c;
+	}
+	else if (v11 == MapType_t::Night)
+	{
+		v92 = (*xadataclrd0dat.colorPalette_var28)[4095];
+		v93 = (*xadataclrd0dat.colorPalette_var28)[4095];
+		v91 = 0xe8;//0xe8
+		v90 = 0x84;//0x84
+	}
+	else if (v11 == MapType_t::Cave)
+	{
+		v92 = (*xadataclrd0dat.colorPalette_var28)[4095];
+		v93 = (*xadataclrd0dat.colorPalette_var28)[4095];
+		v91 = 0x1c;
+		v90 = (*xadataclrd0dat.colorPalette_var28)[240];
+	}
+	if (x_WORD_180660_VGA_type_resolution & 1)
+	{
+		v8 = x >> 1;
+		v9 = y >> 1;
+		scaling *= 2;
+		width >>= 1;
+		height >>= 1;
+	}
+
+	v84 = v9 * screenWidth_18062C + pdwScreenBuffer_351628 + v8;
+	LODWORD(v12) = 0x10000;
+	HIDWORD(v12) = 0x10000 >> 31;
+	v13 = v12 / scaling;
+	v82 = width / 2;
+	v14 = v13 * Maths::sin_DB750[yaw & 0x7FF];
+	v76 = height / 2;
+	v15 = (x_DWORD)Maths::sin_DB750[0x200 + yaw & 0x7FF] * v13;
+	v86 = -v14 >> 16;
+	GetFont_6FC50(x_BYTE_D419D_fonttype);
+	v73 = v15 >> 16;
+	v80x = x_DWORD_EA3E4[D41A0_0.array_0x2BDE[D41A0_0.LevelIndex_0xc].word_0x00a_2BE4_11240];
+	v16 = v80x->dword_0xA4_164x->str_611.array_0x333_819x.word[12];
+	if (v16)
+		v83 = x_DWORD_EA3E4[v16]->word_0x2E_46;
+	//v17 = v80x->dword_0xA4_164;
+	if (v80x->dword_0xA4_164x->word_0x3A_58)
+	{
+		v18 = (signed __int16)(v80x->axis_0x4C_76.x - posX);
+		v19 = (signed __int16)(v80x->axis_0x4C_76.y - posY);
+		v20 = v82 + ((v18 * v73 - v19 * v86) >> 16);
+		v21 = ((v73 * v19 + v86 * v18) >> 16) + v76;
+		v22x = x_DWORD_EA3E4[v80x->dword_0xA4_164x->word_0x3A_58];
+		v23 = (signed __int16)(v22x->axis_0x4C_76.x - posX);
+		v24 = (signed __int16)(v22x->axis_0x4C_76.y - posY);
+		v25 = v82 + ((v23 * v73 - v24 * v86) >> 16);
+		v26 = ((v73 * v24 + v86 * v23) >> 16) + v76;
+		v27 = v82 + ((v23 * v73 - v24 * v86) >> 16) - v20;
+		v75 = Maths::sub_7277A_radix_3d(v27 * v27 + (v26 - v21) * (v26 - v21));
+		v72 = (unsigned __int16)Maths::sub_72633_maybe_tan(v25 - v20, v26 - v21);
+		for (i = D41A0_0.array_0x2BDE[D41A0_0.LevelIndex_0xc].dword_0x012_2BE0_11248 & 3;
+			;
+			*(x_BYTE*)(v84 + v20 + (i * Maths::sin_DB750[v72] >> 16) + screenWidth_18062C * v85) = x_BYTE_F6EE0_tablesx[0x4000 + 256
+			* *(unsigned __int8*)(v84 + v20 + (i * Maths::sin_DB750[v72] >> 16) + screenWidth_18062C * v85)
+			+ (unsigned __int8)(*xadataclrd0dat.colorPalette_var28)[4095]])//castle rope
+		{
+			i += 4;
+			if (i > v75)
+				break;
+
+			if (v72 < 0 || v72 > 2560)
+				break;
+
+			v29 = v20 + (i * Maths::sin_DB750[v72] >> 16);
+			//v29 = v20 + (i * Maths::x_DWORD_DB750ar_ret(4*v72) >> 16);
+
+			if (v72 < 0 || 0x200 + v72 > 2560)
+				break;
+
+			v85 = v21 + (-i * Maths::sin_DB750[0x200 + v72] >> 16);
+
+			if ((1 + 2 * v85) < 0 || (1 + 2 * v85) > 1664 * 8)
+				break;
+
+			if (v29 < 0 || v29 >= width || v85 < 0 || v85 >= height || v29 < x_WORD_F4960[1 + 2 * v85] || v29 >= x_WORD_F4960[2 * v85])
+				break;
+		}
+	}
+	//for all entites
+	//VGA_Debug_Blit(640, 480, pdwScreenBuffer_351628);
+	//for (jy = &D41A0_BYTESTR_0.struct_0x6E8E[1]; jy <= &D41A0_BYTESTR_0.struct_0x6E8E[0x3e8-1]; jy += sizeof(type_str_0x6E8E))
+	for (jy = &D41A0_0.struct_0x6E8E[1]; jy <= &D41A0_0.struct_0x6E8E[0x3e8 - 1]; jy++)
+	{
+		//a3 //adress 242cf9 20-ok 80-ok e0-ok e1-ok e2-ok
+
+		v31 = 0;
+		v89 = 0;
+		v78 = 1;
+		v32 = jy->class_0x3F_63 - 2;//entity type
+		v88 = 0;
+		if (v32 <= 0xDu)
+		{
+			switch (v32)
+			{
+			case 0u:
+				v42 = jy->model_0x40_64;//entity subtype
+				if (v42 > 3u)
+				{
+				LABEL_78:
+					v38 = v90;
+					goto LABEL_121;
+				}
+				switch (v42)
+				{
+				case 0u:
+					if (jy->state_0x45_69 != 2)
+						goto LABEL_78;
+					v94 = v90;
+					v31 = 0;
+					break;
+				case 1u:
+					v31 = 1;
+					if (x_D41A0_BYTEARRAY_4_struct.byteindex_121[3])
+						v94 = (*xadataclrd0dat.colorPalette_var28)[0x88];
+					else
+						v89 = 1;
+					v88 = 27;
+					break;
+				case 2u:
+					if (x_D41A0_BYTEARRAY_4_struct.byteindex_121[2])
+						v43 = (*xadataclrd0dat.colorPalette_var28)[0x888];
+					else
+						v43 = v90;
+					v94 = v43;
+					v31 = 1;
+					v88 = 22;
+					break;
+				case 3u:
+					goto LABEL_123;
+				}
+				break;
+			case 1u:
+				v50 = jy->model_0x40_64;
+				if (v50 >= 2u)
+				{
+					if (v50 <= 2u)
+					{
+						LOBYTE(v51) = jy->id_0x1A_26 != v80x->id_0x1A_26;
+						HIBYTE(v51) = 0;
+						v88 = v51 + 24;
+						v31 = sub_616D0(x_DWORD_EA3E4[jy->id_0x1A_26]->dword_0xA4_164x->word_0x38_56)//castle rope
+							+ 58;
+					}
+					else if (v50 == 3)
+					{
+						if (jy->id_0x1A_26 == v80x->id_0x1A_26 || v83)
+							v31 = sub_616D0(x_DWORD_EA3E4[jy->id_0x1A_26]->dword_0xA4_164x->word_0x38_56)
+							+ 66;
+						if (jy->id_0x1A_26 == v80x->id_0x1A_26)
+						{
+							v88 = 23;
+						}
+						else if (v83)
+						{
+							v88 = 26;
+						}
+					}
+				}
+				break;
+			case 2u:
+			case 4u:
+			case 5u:
+			case 6u:
+			case 0xBu:
+				break;
+			case 3u:
+				v44 = jy->state_0x45_69;
+				if (v44 >= 0xB4u && (v44 <= 0xB4u || v44 >= 0xE8u && v44 <= 0xEAu) || jy->struct_byte_0xc_12_15.byte[0] & 1)
+					break;
+				v45 = jy->id_0x1A_26;
+				if (v45 != jy - D41A0_0.struct_0x6E8E)
+				{
+					v47 = x_BYTE_E88E0x[1 + 3 * GetTrueWizardNumber_61790(x_DWORD_EA3E4[v45]->dword_0xA4_164x->word_0x38_56)];
+					goto LABEL_102;
+				}
+				v46 = jy->model_0x40_64;
+				if (v46 >= 0xCu)
+				{
+					if (v46 <= 0xEu)
+					{
+						v47 = (*xadataclrd0dat.colorPalette_var28)[15];
+					LABEL_102:
+						v94 = v47;
+						goto LABEL_122;
+					}
+					if (v46 == 22)
+					{
+						if (jy->word_0x94_148 != v80x->id_0x1A_26)
+							v88 = 18;
+						v47 = v92;
+						goto LABEL_102;
+					}
+				}
+				v48 = jy->StageVar2_0x49_73;
+				v94 = v92;
+				v49 = 1;
+				if ((v48 == 14 || v48 == 13) && jy->parentId_0x28_40 == v80x->id_0x1A_26)
+					v49 = 0;
+				if (v49)
+					v88 = 19;
+				goto LABEL_122;
+			case 7u:
+				goto LABEL_56;
+			case 8u:
+				v34 = jy->model_0x40_64;
+				if (v34 < 0x27u)
+				{
+					if (v34 >= 0x12u)
+					{
+						if (v34 <= 0x12u)
+							break;
+						if (v34 == 34)
+							v78 = 2;
+					}
+					goto LABEL_56;
+				}
+				if (v34 <= 0x27u)
+					goto LABEL_173;
+				if (v34 < 0x39u)
+				{
+					if (v34 != 45)
+						goto LABEL_56;
+					v40 = jy->word_0x94_148;
+					if (v40)
+					{
+						v41x = x_DWORD_EA3E4[v40];
+						if (v41x->class_0x3F_63 == 3)
+						{
+							if (x_D41A0_BYTEARRAY_4_struct.byteindex_121[3])
+								v38 = x_BYTE_E88E0x[3 * GetTrueWizardNumber_61790(v41x->dword_0xA4_164x->word_0x38_56)];
+							else
+								v38 = x_BYTE_E88E0x[1 + 3 * GetTrueWizardNumber_61790(v41x->dword_0xA4_164x->word_0x38_56)];
+							goto LABEL_121;
+						}
+					}
+					if (!(str_D93C0_bldgprmbuffer[jy->byte_0x46_70].byte_2 & 8))
+						v88 = 21;
+				LABEL_56:
+					v39x = x_DWORD_EA3E4[jy->id_0x1A_26];
+					if (v39x->class_0x3F_63 == 3)
+						v38 = x_BYTE_E88E0x[3 * GetTrueWizardNumber_61790(v39x->dword_0xA4_164x->word_0x38_56)];
+					else
+						v38 = (*xadataclrd0dat.colorPalette_var28)[0xf0f];
+				LABEL_121:
+					v94 = v38;
+					goto LABEL_122;
+				}
+				if (v34 <= 0x39u)
+				{
+				LABEL_173:
+					v35 = jy->word_0x94_148;
+					if (v35 && (v36x = x_DWORD_EA3E4[v35], v36x->class_0x3F_63 == 3))
+					{
+						if (x_D41A0_BYTEARRAY_4_struct.byteindex_121[3])
+							v37 = x_BYTE_E88E0x[3 * GetTrueWizardNumber_61790(v36x->dword_0xA4_164x->word_0x38_56)];
+						else
+							v37 = x_BYTE_E88E0x[1 + 3 * GetTrueWizardNumber_61790(v36x->dword_0xA4_164x->word_0x38_56)];
+						v94 = v37;
+					}
+					else
+					{
+						v94 = v91;
+						v88 = 18;
+					}
+				LABEL_122:
+					v31 = 1;
+				}
+				else
+				{
+					if (v34 < 0x4Eu)
+						goto LABEL_56;
+					if (v34 <= 0x4Eu)
+					{
+						if (jy->word_0x32_50 != D41A0_0.array_0x2BDE[D41A0_0.LevelIndex_0xc].word_0x00a_2BE4_11240)
+							break;
+						v38 = x_BYTE_E88E0x[3 * GetTrueWizardNumber_61790(v80x->dword_0xA4_164x->word_0x38_56)
+							+ x_D41A0_BYTEARRAY_4_struct.byteindex_121[3]];
+						goto LABEL_121;
+					}
+					if (v34 < 0x56u || v34 > 0x57u)
+						goto LABEL_56;
+				}
+				break;
+			case 9u:
+				v33 = jy->model_0x40_64;
+				if (v33 <= 0x2Cu)
+				{
+					switch (v33)
+					{
+					case 0u:
+					case 1u:
+					case 2u:
+					case 3u:
+					case 4u:
+					case 5u:
+					case 6u:
+					case 7u:
+					case 8u:
+					case 9u:
+					case 0xAu:
+					case 0xBu:
+					case 0xDu:
+					case 0xEu:
+					case 0xFu:
+					case 0x10u:
+					case 0x11u:
+					case 0x12u:
+					case 0x13u:
+					case 0x14u:
+					case 0x15u:
+					case 0x16u:
+					case 0x17u:
+					case 0x18u:
+					case 0x19u:
+					case 0x1Au:
+					case 0x1Bu:
+					case 0x1Cu:
+					case 0x1Du:
+					case 0x1Eu:
+					case 0x20u:
+					case 0x21u:
+					case 0x22u:
+					case 0x23u:
+					case 0x24u:
+					case 0x25u:
+					case 0x26u:
+					case 0x27u:
+					case 0x28u:
+					case 0x29u:
+					case 0x2Au:
+					case 0x2Bu:
+					case 0x2Cu:
+						goto LABEL_123;
+					case 0xCu:
+						v31 = 83;
+						v88 = 83;
+						break;
+					case 0x1Fu:
+						v31 = 84;
+						v88 = 83;
+						break;
+					}
+				}
+				break;
+			case 0xAu:
+			case 0xDu:
+				if (jy->struct_byte_0xc_12_15.byte[0] & 1)
+					break;
+				v38 = (*xadataclrd0dat.colorPalette_var28)[3840];
+				v88 = 20;
+				goto LABEL_121;
+			case 0xCu:
+				if (jy->model_0x40_64 != 5)
+					break;
+				if (x_D41A0_BYTEARRAY_4_struct.byteindex_121[3])
+					v38 = (*xadataclrd0dat.colorPalette_var28)[3840];
+				else
+					v38 = (*xadataclrd0dat.colorPalette_var28)[4095];
+				goto LABEL_121;
+			}
+		}
+	LABEL_123:
+		if (v31)//adress 243207
+		{
+			v52 = (int16_t)(jy->axis_0x4C_76.x - posX);
+			v53 = (int16_t)(jy->axis_0x4C_76.y - posY);
+			v54 = v82 + ((v52 * v73 - v53 * v86) >> 16);
+			if (v54 >= 0 && v54 < width)
+			{//24324e
+				v55 = ((v73 * v53 + v86 * v52) >> 16) + v76;
+				if (v55 >= 0 && v55 < height && v54 >= x_WORD_F4960[1 + 2 * v55] && v54 < x_WORD_F4960[2 * v55])
+				{//adress 24329e 0x1e xx 0x17
+					v56 = (char*)(v54 + v84 + v55 * screenWidth_18062C);
+					if (!v89)
+					{
+						if (v31 <= 1u)
+						{
+							v56[0] = v94;//here //adress 243429
+							//VGA_Debug_Blit(640, 480, pdwScreenBuffer_351628);
+							if (v78 > 1)
+							{
+								v57 = screenWidth_18062C;
+								v56[1] = v94;
+								v58 = &v56[v57];
+								v58[0] = v94;
+								v58[1] = v94;
+							}
+						}
+						else if (x_WORD_180660_VGA_type_resolution == 1)
+						{
+							if (v31 < 0x42u || v31 >= 0x4Au)
+							{
+								if (v31 < 0x3Au || v31 >= 0x42u)
+								{
+									if (v31 == 83 || v31 == 84)
+										sub_2BB40_draw_bitmap(
+											2 * v54 - (*filearray_2aa18c[filearrayindex_MSPRD00DATTAB].posistruct)[v31].width_4 / 2,
+											2 * v55 - (*filearray_2aa18c[filearrayindex_MSPRD00DATTAB].posistruct)[v31].height_5 / 2,
+											(*filearray_2aa18c[filearrayindex_MSPRD00DATTAB].posistruct)[v31]);
+								}
+								else
+								{
+									//sub_2BB40_draw_bitmap(2 * v54, 2 * v55 - *(unsigned __int8 *)(6 * v31 + **filearray_2aa18c[6] + 5), (uint8_t**)(6 * v31 + **filearray_2aa18c[6]));
+									sub_2BB40_draw_bitmap(2 * v54, 2 * v55 - (*filearray_2aa18c[filearrayindex_MSPRD00DATTAB].posistruct)[v31].height_5, (*filearray_2aa18c[filearrayindex_MSPRD00DATTAB].posistruct)[v31]);
+								}
+							}
+							else
+							{
+								sub_2BB40_draw_bitmap(
+									2 * v54 - (*filearray_2aa18c[filearrayindex_MSPRD00DATTAB].posistruct)[v31].width_4 / 2,
+									2 * v55 - (*filearray_2aa18c[filearrayindex_MSPRD00DATTAB].posistruct)[v31].height_5,
+									(*filearray_2aa18c[filearrayindex_MSPRD00DATTAB].posistruct)[v31]);
+							}
+						}
+						else if (v31 >= 0x3Cu)
+						{
+							sub_2BB40_draw_bitmap(
+								v54 - (*filearray_2aa18c[filearrayindex_MSPRD00DATTAB].posistruct)[v31].width_4 / 2,
+								v55 - (*filearray_2aa18c[filearrayindex_MSPRD00DATTAB].posistruct)[v31].height_5,
+								(*filearray_2aa18c[filearrayindex_MSPRD00DATTAB].posistruct)[v31]);
+						}
+						else
+						{
+							//sub_2BB40_draw_bitmap(v54, v55 - *(unsigned __int8 *)(6 * v31 + **filearray_2aa18c[6] + 5), (uint8_t**)(6 * v31 + **filearray_2aa18c[6]));
+							sub_2BB40_draw_bitmap(v54, v55 - (*filearray_2aa18c[filearrayindex_MSPRD00DATTAB].posistruct)[v31].height_5, (*filearray_2aa18c[filearrayindex_MSPRD00DATTAB].posistruct)[v31]);
+						}
+					}
+					if (v88)
+						sub_885E0(jy, (signed __int16)v54, (signed __int16)v55, v88);
+				}
+			}
+		}
+	}
+
+	if (v83)//player cross or other
+	{
+		v87 = 0;
+		v74x = x_DWORD_EA3E4[D41A0_0.array_0x2BDE[D41A0_0.LevelIndex_0xc].word_0x00a_2BE4_11240];
+		while (v87 < D41A0_0.word_0xe)
+		{
+			if (v87 != D41A0_0.LevelIndex_0xc)
+			{
+				//v79 = 2124 * v87;
+				v59 = D41A0_0.array_0x2BDE[v87].word_0x00a_2BE4_11240;
+				v60x = x_DWORD_EA3E4[v59];
+				if (v60x->life_0x8 >= 0)
+				{
+					if (sub_63570(v74x, x_DWORD_EA3E4[v59]))
+					{
+						v61 = (signed __int16)(v60x->axis_0x4C_76.x - posX);
+						v62 = (signed __int16)(v60x->axis_0x4C_76.y - posY);
+						v63 = v82 + ((v61 * v73 - v62 * v86) >> 16);
+						if (v63 >= 0 && v63 < width)
+						{
+							v64 = v76 + ((v86 * v61 + v73 * v62) >> 16);
+							if (v64 >= 0 && v64 < height && v63 >= x_WORD_F4960[1 + 2 * v64] && v63 < x_WORD_F4960[2 * v64])
+							{
+								v65 = (unsigned __int8*)(v64 * screenWidth_18062C + v63 + v84);
+								v66 = x_BYTE_E88E0x[3 * GetTrueWizardNumber_61790(v60x->dword_0xA4_164x->word_0x38_56)];
+								*v65 = v66;
+								if (x_WORD_180660_VGA_type_resolution == 1)
+									DrawText_2BC10(D41A0_0.array_0x2BDE[v87].WizardName_0x39f_2BFA_12157, 2 * v63 + 4, 2 * v64, v66);
+								else
+									DrawText_2BC10(D41A0_0.array_0x2BDE[v87].WizardName_0x39f_2BFA_12157, v63 + 4, v64, v66);
+							}
+						}
+					}
+				}
+			}
+			v87++;
+		}
+	}
+	v67 = width / 12;
+	v68 = (x_BYTE*)(screenWidth_18062C * (int)(height / 2) + width / 2 + v84 - 1);
+	index = (char*)&x_BYTE_F6EE0_tablesx[0x4000 + 256 * v93];
+	v70 = (x_BYTE*)(screenWidth_18062C * (int)(height / 2) + width / 2 + v84 - 1);
+	v77 = (x_BYTE*)(screenWidth_18062C * (int)(height / 2) + width / 2 + v84 - 1);
+	v81 = (x_BYTE*)(screenWidth_18062C * (int)(height / 2) + width / 2 + v84 - 1);
+	for (*v68 = index[(unsigned __int8)*v68]; v67; *v68 = x_BYTE_F6EE0_tablesx[0x4000 + 256 * v93 + v71])
+	{
+		v81 -= screenWidth_18062C;
+		v70 += screenWidth_18062C;
+		v77++;
+		*v81 = x_BYTE_F6EE0_tablesx[0x4000 + 256 * v93 + (unsigned __int8)*v81];
+		*v77 = x_BYTE_F6EE0_tablesx[0x4000 + 256 * v93 + (unsigned __int8)*v77];
+		*v70 = x_BYTE_F6EE0_tablesx[0x4000 + 256 * v93 + (unsigned __int8)*v70];
+		v71 = (unsigned __int8)*(v68-- - 1);
+		v67--;
+	}
+}
+
+//----- (000627F0) --------------------------------------------------------
+void sub_627F0_draw_minimap_entites_a(int16_t x, int16_t y, int16_t posX, int16_t posY, uint16_t width, uint16_t height, int16_t yaw, int16_t scaling)//2437f0
+{//
+	int v8; // edx
+	int v9; // esi
+	int v10; // ebx
+	MapType_t v11; // al
+	char v12; // al
+	__int64 v13; // rtt
+	int v14; // ebx
+	int v15; // edx
+	int v16; // ebx
+	__int16 v17; // dx
+	//int v18; // ebx
+	int v19; // edx
+	int v20; // eax
+	int v21; // esi
+	int v22; // edx
+	type_event_0x6E8E* v23x; // eax
+	int v24; // edi
+	int v25; // edx
+	int v26; // eax
+	int v27; // ebx
+	int v28; // ST48_4
+	int v29; // edx
+	int v30; // eax
+	int v31; // ebx
+	uint8_t* v32x; // edx
+	uint8_t* v33x; // ST44_4
+	char v34; // bl
+	type_event_0x6E8E* iy; // ebx
+	unsigned __int8 v36; // al
+	unsigned __int8 v37; // ah
+	unsigned __int8 v38; // ah
+	unsigned __int8 v39; // ah
+	unsigned __int16 v40; // si
+	type_event_0x6E8E* v41x; // eax
+	char v42; // al
+	char v43; // al
+	type_event_0x6E8E* v44x; // eax
+	unsigned __int16 v45; // si
+	type_event_0x6E8E* v46x; // eax
+	//bool v47; // zf
+	unsigned __int8 v48; // ah
+	unsigned __int8 v49; // ah
+	int v50; // eax
+	unsigned __int8 v51; // al
+	char v52; // al
+	char v53; // al
+	char v54; // ah
+	unsigned __int8 v55; // ah
+	__int16 v56; // ax
+	int v57; // edx
+	int v58; // ecx
+	int v59; // esi
+	int v60; // edi
+	uint8_t* v61x; // ecx
+	uint8_t* v62x; // edx
+	uint8_t* v63; // edx
+	uint8_t* v64; // ecx
+	int v65; // eax
+	type_event_0x6E8E* v66x; // ebx
+	int v67; // edx
+	int v68; // eax
+	int v69; // esi
+	int v70; // edi
+	uint8_t* v71; // ST2C_4
+	uint8_t* v72; // ST28_4
+	unsigned __int8 v73; // al
+	int v74; // eax
+	uint8_t* v75; // esi
+	uint8_t* v76; // ebx
+	char* index; // eax
+	uint8_t* v78; // edx
+	char v79; // cl
+	char v80; // cl
+	char v81; // cl
+	char v82; // cl
+	uint8_t* v83; // [esp+0h] [ebp-1Eh]
+	uint8_t* v84; // [esp+4h] [ebp-1Ah]
+	int v85; // [esp+8h] [ebp-16h]
+	int v86; // [esp+Ch] [ebp-12h]
+	int v87; // [esp+10h] [ebp-Eh]
+	uint8_t* v88; // [esp+14h] [ebp-Ah]
+	int v89; // [esp+20h] [ebp+2h]
+	type_event_0x6E8E* v90x; // [esp+24h] [ebp+6h]
+	signed int v91; // [esp+28h] [ebp+Ah]
+	signed int v92; // [esp+2Ch] [ebp+Eh]
+	int v93; // [esp+30h] [ebp+12h]
+	int v94; // [esp+38h] [ebp+1Ah]
+	uint8_t* v95; // [esp+3Ch] [ebp+1Eh]
+	//int v96; // [esp+40h] [ebp+22h]
+	int v97; // [esp+44h] [ebp+26h]
+	int v98; // [esp+48h] [ebp+2Ah]
+	uint8_t* v99x=0; // [esp+4Ch] [ebp+2Eh]
+	uint8_t* v100x=0; // [esp+50h] [ebp+32h]
+	uint8_t* v101; // [esp+54h] [ebp+36h]
+	type_event_0x6E8E* v102x; // [esp+58h] [ebp+3Ah]
+	unsigned __int16 j; // [esp+5Ch] [ebp+3Eh]
+	unsigned __int16 v104; // [esp+60h] [ebp+42h]
+	char v105; // [esp+68h] [ebp+4Ah]
+	unsigned __int8 v106; // [esp+6Ch] [ebp+4Eh]
+	char v107; // [esp+70h] [ebp+52h]
+	char v108; // [esp+74h] [ebp+56h]
+	char v109; // [esp+78h] [ebp+5Ah]
+	char v110; // [esp+7Ch] [ebp+5Eh]
+
+	v8 = x;
+	v9 = y;
+	v10 = scaling;
+	v11 = D41A0_0.terrain_2FECE.MapType;
+	v93 = 0;
+	if (v11 == MapType_t::Day)
+	{
+    v109 = (*xadataclrd0dat.colorPalette_var28)[0x000];
+    v106 = (*xadataclrd0dat.colorPalette_var28)[0x000];
+    v108 = 28;
+    v107 = -24;
+	}
+	else if (v11 == MapType_t::Night)
+	{
+		v109 = (*xadataclrd0dat.colorPalette_var28)[4095];
+		v106 = (*xadataclrd0dat.colorPalette_var28)[4095];
+		v12 = -124;
+		v107 = -24;
+	  v108 = v12;
+	}
+	else // Cave
+	{
+		v109 = (*xadataclrd0dat.colorPalette_var28)[4095];
+		v106 = (*xadataclrd0dat.colorPalette_var28)[4095];
+		v107 = 28;
+		v12 = (*xadataclrd0dat.colorPalette_var28)[240];
+	  v108 = v12;
+	}
+
+	if (x_WORD_180660_VGA_type_resolution & 1)
+	{
+		v8 = x >> 1;
+		v9 = y >> 1;
+		v10 = 2 * scaling;
+		width >>= 1;
+		height >>= 1;
+	}
+	//v99 = v9 * x_DWORD_18062C_resolution_x + (int)pdwScreenBuffer + v8;
+	v99x = &pdwScreenBuffer_351628[v9 * screenWidth_18062C + v8];
+	v100x = &x_DWORD_E9C3C[v9 * screenWidth_18062C + v8];
+	LODWORD(v13) = 0x10000;
+	HIDWORD(v13) = 0x10000 >> 31;
+	v14 = v13 / v10;
+	v98 = width / 2;
+	v15 = v14 * Maths::sin_DB750[yaw & 0x7FF];
+	v16 = (x_DWORD)Maths::sin_DB750[0x200 + yaw & 0x7FF] * v14;
+	v85 = height / 2;
+	v86 = -v15 >> 16;
+	GetFont_6FC50(x_BYTE_D419D_fonttype);
+	v87 = v16 >> 16;
+	v102x = x_DWORD_EA3E4[D41A0_0.array_0x2BDE[D41A0_0.LevelIndex_0xc].word_0x00a_2BE4_11240];
+	v17 = v102x->dword_0xA4_164x->str_611.array_0x333_819x.word[12];
+	if (v17)
+		v93 = x_DWORD_EA3E4[v17]->word_0x2E_46;
+	//v18 = v102x->dword_0xA4_164;
+	if (v102x->dword_0xA4_164x->word_0x3A_58)
+	{
+		v19 = (signed __int16)(v102x->axis_0x4C_76.x - posX);
+		v20 = (signed __int16)(v102x->axis_0x4C_76.y - posY);
+		v21 = v98 + ((v19 * v87 - v20 * v86) >> 16);
+		v22 = v87 * v20 + v86 * v19;
+		v23x = x_DWORD_EA3E4[v102x->dword_0xA4_164x->word_0x3A_58];
+		v24 = (v22 >> 16) + v85;
+		v25 = (signed __int16)(v23x->axis_0x4C_76.x - posX);
+		v26 = (signed __int16)(v23x->axis_0x4C_76.y - posY);
+		v27 = v98 + ((v25 * v87 - v26 * v86) >> 16);
+		v28 = ((v86 * v25 + v87 * v26) >> 16) + v85;
+		v29 = v98 + ((v25 * v87 - v26 * v86) >> 16) - v21;
+		v91 = Maths::sub_7277A_radix_3d(v29 * v29 + (v28 - v24) * (v28 - v24));
+		v89 = (unsigned __int16)Maths::sub_72633_maybe_tan(v27 - v21, v28 - v24);
+		v30 = D41A0_0.array_0x2BDE[D41A0_0.LevelIndex_0xc].dword_0x012_2BE0_11248 & 3;
+		while (1)
+		{
+			v30 += 4;
+			if (v30 > v91)
+				break;
+			v31 = v21 + (v30 * Maths::sin_DB750[v89] >> 16);
+			v94 = v24 + (-v30 * (signed int)Maths::sin_DB750[0x200 + v89] >> 16);
+			if (v31 < 0 || v31 >= width || v94 < 0 || v94 >= height || v31 < x_WORD_F4960[1 + 2 * v94] || v31 >= x_WORD_F4960[2 * v94])
+				break;
+			v32x = &v99x[screenWidth_18062C * v94 + v31];
+			v33x = &v100x[screenWidth_18062C * v94 + v31]; //(char*)(v100 + v31 + x_DWORD_18062C_resolution_x * v94);
+			v34 = x_BYTE_F6EE0_tablesx[0x4000 + 256 * *v32x + (unsigned __int8)(*xadataclrd0dat.colorPalette_var28)[4095]];
+			*v33x = v34;
+			*v32x = v34;
+		}
+	}
+	for (iy = &D41A0_0.struct_0x6E8E[1]; iy <= &D41A0_0.struct_0x6E8E[0x3e8 - 1]; iy++)
+	{
+		v36 = 0;
+		v105 = 0;
+		v92 = 1;
+		v37 = iy->class_0x3F_63 - 2;
+		v104 = 0;
+		if (v37 <= 0xDu)
+		{
+			switch (v37)
+			{
+			case 0u:
+				v48 = iy->model_0x40_64;
+				if (v48 > 3u)
+				{
+				LABEL_74:
+					v110 = v108;
+					goto LABEL_118;
+				}
+				switch (v48)
+				{
+				case 0u:
+					if (iy->state_0x45_69 != 2)
+						goto LABEL_74;
+					v110 = v108;
+					v36 = 0;
+					break;
+				case 1u:
+					v36 = 1;
+					if (x_D41A0_BYTEARRAY_4_struct.byteindex_121[2])
+						v110 = (*xadataclrd0dat.colorPalette_var28)[0x88];
+					else
+						v105 = 1;
+					v104 = 27;
+					break;
+				case 2u:
+					v110 = v108;
+					v104 = 22;
+					goto LABEL_118;
+				case 3u:
+					goto LABEL_119;
+				}
+				break;
+			case 1u:
+				v55 = iy->model_0x40_64;
+				if (v55 >= 2u)
+				{
+					if (v55 <= 2u)
+					{
+						LOBYTE(v56) = iy->id_0x1A_26 != v102x->id_0x1A_26;
+						HIBYTE(v56) = 0;
+						v104 = v56 + 24;
+						v36 = x_DWORD_EA3E4[iy->id_0x1A_26]->dword_0xA4_164x->word_0x38_56 + 58;
+					}
+					else if (v55 == 3)
+					{
+						if (iy->id_0x1A_26 == v102x->id_0x1A_26 || v93)
+							v36 = x_DWORD_EA3E4[iy->id_0x1A_26]->dword_0xA4_164x->word_0x38_56 + 66;
+						if (iy->id_0x1A_26 == v102x->id_0x1A_26)
+						{
+							v104 = 23;
+						}
+						else if (v93)
+						{
+							v104 = 26;
+						}
+					}
+				}
+				break;
+			case 2u:
+			case 4u:
+			case 5u:
+			case 6u:
+			case 0xBu:
+				break;
+			case 3u:
+				v49 = iy->state_0x45_69;
+				if (v49 >= 0xB4u && (v49 <= 0xB4u || v49 >= 0xE8u && v49 <= 0xEAu) || iy->struct_byte_0xc_12_15.byte[0] & 1)
+					break;
+				v50 = iy->id_0x1A_26;
+				if (v50 != iy - D41A0_0.struct_0x6E8E)
+				{
+					v52 = x_BYTE_E88E0x[1 + 3 * GetTrueWizardNumber_61790(x_DWORD_EA3E4[v50]->dword_0xA4_164x->word_0x38_56)];
+					goto LABEL_98;
+				}
+				v51 = iy->model_0x40_64;
+				if (v51 >= 0xCu)
+				{
+					if (v51 <= 0xEu)
+					{
+						v52 = (*xadataclrd0dat.colorPalette_var28)[15];
+					LABEL_98:
+						v110 = v52;
+						goto LABEL_118;
+					}
+					if (v51 == 22)
+					{
+						if (iy->word_0x94_148 != v102x->id_0x1A_26)
+							v104 = 18;
+						v52 = v109;
+						goto LABEL_98;
+					}
+				}
+				v110 = v109;
+				v53 = 1;
+				v54 = iy->StageVar2_0x49_73;
+				if ((v54 == 14 || v54 == 13) && iy->parentId_0x28_40 == v102x->id_0x1A_26)
+					v53 = 0;
+				if (v53)
+					v104 = 19;
+				goto LABEL_118;
+			case 7u:
+				goto LABEL_57;
+			case 8u:
+				v39 = iy->model_0x40_64;
+				if (v39 < 0x27u)
+				{
+					if (v39 >= 0x12u)
+					{
+						if (v39 <= 0x12u)
+							break;
+						if (v39 == 34)
+							v92 = 2;
+					}
+					goto LABEL_57;
+				}
+				if (v39 <= 0x27u)
+					goto LABEL_169;
+				if (v39 < 0x39u)
+				{
+					if (v39 == 45)
+					{
+						v45 = iy->word_0x94_148;
+						if (v45 && (v46x = x_DWORD_EA3E4[v45], v46x->class_0x3F_63 == 3))
+						{
+							//v47 = x_D41A0_BYTEARRAY_4_struct.byteindex_121[3] == 0;
+							GetTrueWizardNumber_61790(v46x->dword_0xA4_164x->word_0x38_56);
+						}
+						else
+						{
+							if (!(str_D93C0_bldgprmbuffer[iy->byte_0x46_70].byte_2 & 8))
+								v104 = 21;
+							v105 = 1;
+						}
+					}
+				LABEL_57:
+					v44x = x_DWORD_EA3E4[iy->id_0x1A_26];
+					if (v44x->class_0x3F_63 == 3)
+						v43 = x_BYTE_E88E0x[3 * GetTrueWizardNumber_61790(v44x->dword_0xA4_164x->word_0x38_56)];
+					else
+						v43 = (*xadataclrd0dat.colorPalette_var28)[0xf0f];
+				LABEL_117:
+					v110 = v43;
+					goto LABEL_118;
+				}
+				if (v39 <= 0x39u)
+				{
+				LABEL_169:
+					v40 = iy->word_0x94_148;
+					if (v40 && (v41x = x_DWORD_EA3E4[v40], v41x->class_0x3F_63 == 3))
+					{
+						if (x_D41A0_BYTEARRAY_4_struct.byteindex_121[3])
+							v42 = x_BYTE_E88E0x[3 * GetTrueWizardNumber_61790(v41x->dword_0xA4_164x->word_0x38_56)];
+						else
+							v42 = x_BYTE_E88E0x[1 + 3 * GetTrueWizardNumber_61790(v41x->dword_0xA4_164x->word_0x38_56)];
+						v110 = v42;
+					}
+					else
+					{
+						v110 = v107;
+						v104 = 18;
+					}
+				LABEL_118:
+					v36 = 1;
+				}
+				else
+				{
+					if (v39 < 0x4Eu)
+						goto LABEL_57;
+					if (v39 <= 0x4Eu)
+					{
+						if (iy->word_0x32_50 != D41A0_0.array_0x2BDE[D41A0_0.LevelIndex_0xc].word_0x00a_2BE4_11240)
+							break;
+						v43 = x_BYTE_E88E0x[3 * GetTrueWizardNumber_61790(v102x->dword_0xA4_164x->word_0x38_56)
+							+ x_D41A0_BYTEARRAY_4_struct.byteindex_121[3]];
+						goto LABEL_117;
+					}
+					if (v39 < 0x56u || v39 > 0x57u)
+						goto LABEL_57;
+				}
+				break;
+			case 9u:
+				v38 = iy->model_0x40_64;
+				if (v38 <= 0x2Cu)
+				{
+					switch (v38)
+					{
+					case 0u:
+					case 1u:
+					case 2u:
+					case 3u:
+					case 4u:
+					case 5u:
+					case 6u:
+					case 7u:
+					case 8u:
+					case 9u:
+					case 0xAu:
+					case 0xBu:
+					case 0xDu:
+					case 0xEu:
+					case 0xFu:
+					case 0x10u:
+					case 0x11u:
+					case 0x12u:
+					case 0x13u:
+					case 0x14u:
+					case 0x15u:
+					case 0x16u:
+					case 0x17u:
+					case 0x18u:
+					case 0x19u:
+					case 0x1Au:
+					case 0x1Bu:
+					case 0x1Cu:
+					case 0x1Du:
+					case 0x1Eu:
+					case 0x20u:
+					case 0x21u:
+					case 0x22u:
+					case 0x23u:
+					case 0x24u:
+					case 0x25u:
+					case 0x26u:
+					case 0x27u:
+					case 0x28u:
+					case 0x29u:
+					case 0x2Au:
+					case 0x2Bu:
+					case 0x2Cu:
+						goto LABEL_119;
+					case 0xCu:
+						v36 = 83;
+						v104 = 83;
+						break;
+					case 0x1Fu:
+						v36 = 84;
+						v104 = 83;
+						break;
+					}
+				}
+				break;
+			case 0xAu:
+			case 0xDu:
+				if (iy->struct_byte_0xc_12_15.byte[0] & 1)
+					break;
+				v43 = (*xadataclrd0dat.colorPalette_var28)[3840];
+				v104 = 20;
+				goto LABEL_117;
+			case 0xCu:
+				if (iy->model_0x40_64 != 5)
+					break;
+				if (x_D41A0_BYTEARRAY_4_struct.byteindex_121[3])
+					v43 = (*xadataclrd0dat.colorPalette_var28)[3840];
+				else
+					v43 = (*xadataclrd0dat.colorPalette_var28)[4095];
+				goto LABEL_117;
+			}
+		}
+	LABEL_119:
+		if (v36)
+		{
+			v57 = (signed __int16)(iy->axis_0x4C_76.x - posX);
+			v58 = (signed __int16)(iy->axis_0x4C_76.y - posY);
+			v59 = v98 + ((v57 * v87 - v58 * v86) >> 16);
+			if (v59 >= 0 && v59 < width)
+			{
+				v60 = ((v87 * v58 + v86 * v57) >> 16) + v85;
+				if (v60 >= 0 && v60 < height && v59 >= x_WORD_F4960[1 + 2 * v60] && v59 < x_WORD_F4960[2 * v60])
+				{
+					/*v61 = (char*)(v60 * x_DWORD_18062C_resolution_x + v59 + v99);
+					v62 = (char*)(v60 * x_DWORD_18062C_resolution_x + v59 + v100);*/
+					v61x = &v99x[v60 * screenWidth_18062C + v59];
+					v62x = &v100x[v60 * screenWidth_18062C + v59];
+					if (!v105)
+					{
+						if (v36 <= 1u)
+						{
+							*v62x = v110;
+							*v61x = v110;
+							if (v92 > 1)
+							{
+								v62x[1] = v110;
+								v61x[1] = v110;
+								v63 = &v62x[screenWidth_18062C];
+								*v63 = v110;
+								v64 = &v61x[screenWidth_18062C];
+								*v64 = v110;
+								v63[1] = v110;
+								v64[1] = v110;
+							}
+						}
+						else if (x_WORD_180660_VGA_type_resolution == 1)
+						{
+							if (v36 < 0x42u || v36 >= 0x4Au)
+							{
+								if (v36 < 0x3Au || v36 >= 0x42u)
+								{
+									/*if ( v36 == 83 || v36 == 84 )
+									  sub_2BB40_draw_bitmap(
+										2 * v59 - *(unsigned __int8 *)(6 * v36 + **filearray_2aa18c[6] + 4) / 2,
+										2 * v60 - *(unsigned __int8 *)(6 * v36 + **filearray_2aa18c[6] + 5) / 2,
+										  (uint8_t**)(6 * v36 + **filearray_2aa18c[6]));*/
+									if (v36 == 83 || v36 == 84)
+										sub_2BB40_draw_bitmap(
+											2 * v59 - (*filearray_2aa18c[filearrayindex_MSPRD00DATTAB].posistruct)[v36].width_4 / 2,
+											2 * v60 - (*filearray_2aa18c[filearrayindex_MSPRD00DATTAB].posistruct)[v36].height_5 / 2,
+											(*filearray_2aa18c[filearrayindex_MSPRD00DATTAB].posistruct)[v36]);
+								}
+								else
+								{
+									//sub_2BB40_draw_bitmap(2 * v59, 2 * v60 - *(unsigned __int8 *)(**filearray_2aa18c[6] + 6 * v36 + 5), (uint8_t**)(**filearray_2aa18c[6] + 6 * v36));
+									sub_2BB40_draw_bitmap(2 * v59, 2 * v60 - (*filearray_2aa18c[filearrayindex_MSPRD00DATTAB].posistruct)[v36].height_5, (*filearray_2aa18c[filearrayindex_MSPRD00DATTAB].posistruct)[v36]);
+								}
+							}
+							else
+							{
+								/*sub_2BB40_draw_bitmap(
+								  2 * v59 - *(unsigned __int8 *)(6 * v36 + **filearray_2aa18c[6] + 4) / 2,
+								  2 * v60 - *(unsigned __int8 *)(6 * v36 + **filearray_2aa18c[6] + 5),
+									(uint8_t**)(6 * v36 + **filearray_2aa18c[6]));*/
+								sub_2BB40_draw_bitmap(
+									2 * v59 - (*filearray_2aa18c[filearrayindex_MSPRD00DATTAB].posistruct)[v36].width_4 / 2,
+									2 * v60 - (*filearray_2aa18c[filearrayindex_MSPRD00DATTAB].posistruct)[v36].height_5,
+									(*filearray_2aa18c[filearrayindex_MSPRD00DATTAB].posistruct)[v36]);
+							}
+						}
+						else if (v36 >= 0x3Cu)
+						{
+							/*sub_2BB40_draw_bitmap(
+							  v59 - *(unsigned __int8 *)(6 * v36 + **filearray_2aa18c[6] + 4) / 2,
+							  v60 - *(unsigned __int8 *)(6 * v36 + **filearray_2aa18c[6] + 5),
+								(uint8_t**)(6 * v36 + **filearray_2aa18c[6]));*/
+							sub_2BB40_draw_bitmap(
+								v59 - (*filearray_2aa18c[filearrayindex_MSPRD00DATTAB].posistruct)[v36].width_4 / 2,
+								v60 - (*filearray_2aa18c[filearrayindex_MSPRD00DATTAB].posistruct)[v36].height_5,
+								(*filearray_2aa18c[filearrayindex_MSPRD00DATTAB].posistruct)[v36]);
+						}
+						else
+						{
+							//sub_2BB40_draw_bitmap(v59, v60 - *(unsigned __int8 *)(6 * v36 + **filearray_2aa18c[6] + 5), (uint8_t**)(6 * v36 + **filearray_2aa18c[6]));
+							sub_2BB40_draw_bitmap(v59, v60 - (*filearray_2aa18c[filearrayindex_MSPRD00DATTAB].posistruct)[v36].height_5, (*filearray_2aa18c[filearrayindex_MSPRD00DATTAB].posistruct)[v36]);
+						}
+					}
+					if (v104)
+						sub_885E0(iy, (signed __int16)v59, (signed __int16)v60, v104);
+				}
+			}
+		}
+	}
+	if (v93)
+	{
+		v90x = x_DWORD_EA3E4[D41A0_0.array_0x2BDE[D41A0_0.LevelIndex_0xc].word_0x00a_2BE4_11240];
+		for (j = 0; j < D41A0_0.word_0xe; j++)
+		{
+			if (j != D41A0_0.LevelIndex_0xc)
+			{
+				//v96 = 2124 * j;
+				v65 = D41A0_0.array_0x2BDE[j].word_0x00a_2BE4_11240;
+				v66x = x_DWORD_EA3E4[v65];
+				if (v66x->life_0x8 >= 0)
+				{
+					if (sub_63570(v90x, x_DWORD_EA3E4[v65]))
+					{
+						v67 = (signed __int16)(v66x->axis_0x4C_76.x - posX);
+						v68 = (signed __int16)(v66x->axis_0x4C_76.y - posY);
+						v69 = v98 + ((v67 * v87 - v68 * v86) >> 16);
+						if (v69 >= 0 && v69 < width)
+						{
+							v70 = ((v86 * v67 + v87 * v68) >> 16) + v85;
+							if (v70 >= 0 && v70 < height && v69 >= x_WORD_F4960[1 + 2 * v70] && v69 < x_WORD_F4960[2 * v70])
+							{
+								v71 = &v99x[v70 * screenWidth_18062C + v69];
+								v72 = &v100x[v70 * screenWidth_18062C + v69];
+								v73 = x_BYTE_E88E0x[1 + 3 * GetTrueWizardNumber_61790(v66x->dword_0xA4_164x->word_0x38_56)];
+								*v72 = v73;
+								*v71 = v73;
+								if (x_WORD_180660_VGA_type_resolution == 1)
+									DrawText_2BC10(D41A0_0.array_0x2BDE[j].WizardName_0x39f_2BFA_12157, 2 * v69 + 4, 2 * v70, v73);
+								else
+									DrawText_2BC10(D41A0_0.array_0x2BDE[j].WizardName_0x39f_2BFA_12157, v69 + 4, v70, v73);
+							}
+						}
+					}
+				}
+			}
+		}
+	}
+	v97 = width / 12;
+	v74 = screenWidth_18062C * (height / 2);
+	v75 = &v100x[v74 + width / 2 - 1];
+	v76 = &v99x[v74 + width / 2 - 1];
+	index = (char*)&x_BYTE_F6EE0_tablesx[0x4000 + 256 * v106];
+	v88 = v76;
+	v101 = v76;
+	*v76 = index[(unsigned __int8)*v76];
+	v78 = v76;
+	v83 = v75;
+	v84 = v75;
+	v95 = v75;
+	for (*v75 = index[(unsigned __int8)*v76]; v97; --v97)
+	{
+		v101++;
+		v88 += screenWidth_18062C;
+		v78 -= screenWidth_18062C;
+		v95 -= screenWidth_18062C;
+		v84++;
+		v83 += screenWidth_18062C;
+		v79 = x_BYTE_F6EE0_tablesx[0x4000 + 256 * v106 + (unsigned __int8)*v78];
+		*v78 = v79;
+		*v95 = v79;
+		v80 = x_BYTE_F6EE0_tablesx[0x4000 + 256 * v106 + (unsigned __int8)*v101];
+		*v101 = v80;
+		*v84 = v80;
+		v81 = x_BYTE_F6EE0_tablesx[0x4000 + 256 * v106 + (unsigned __int8)*v88];
+		*v88 = v81;
+		*v83 = v81;
+		v82 = x_BYTE_F6EE0_tablesx[0x4000 + 256 * v106 + (unsigned __int8)*(v76 - 1)];
+		*(v76 - 1) = v82;
+		*(v75 - 1) = v82;
+		v76--;
+		v75--;
+	}
+	//return result;
+}
+
+//----- (00063570) --------------------------------------------------------
+char sub_63570(type_event_0x6E8E* a1x, type_event_0x6E8E* a2x)//244570
+{
+	type_event_0x6E8E* v2x; // ecx
+	char v3; // ah
+	type_event_0x6E8E* v4x; // ebx
+	char v5; // dl
+	unsigned __int8 v6; // al
+
+	v2x = x_DWORD_EA3E4[a1x->dword_0xA4_164x->str_611.array_0x333_819x.word[12]];
+	v3 = 1;
+	v4x = x_DWORD_EA3E4[a2x->dword_0xA4_164x->str_611.array_0x333_819x.word[4]];
+	v5 = 0;
+	if (v4x > x_DWORD_EA3E4[0] && v4x->word_0x2E_46)
+		v5 = 1;
+	if (v2x > x_DWORD_EA3E4[0])
+	{
+		v6 = v2x->byte_0x46_70;
+		if (v6 < 1u)
+		{
+			if (v6)
+				return v3;
+			if (a2x->dword_0xA4_164x->byte_0x1BF_447)
+				return 0;
+		}
+		else if (v6 > 1u)
+		{
+			return v3;
+		}
+		if (v5)
+			return 0;
+	}
+	return v3;
+}
+// EA3E4: using guessed type int x_DWORD_EA3E4[];
+
+//----- (00063600) --------------------------------------------------------
+void DrawMinimap_63600(int16_t x, int16_t y, int16_t posX, int16_t posY, uint16_t width, uint16_t height, int16_t yaw, int16_t scaling, int a10)//244600
+{
+	if (D41A0_0.m_GameSettings.m_Display.m_uiScreenSize == 1)
+		sub_63670_draw_minimap_a(x, y, posX, posY, width, height, yaw, scaling, a10);
+	else
+		sub_63C90_draw_minimap_b(x, y, posX, posY, width, height, yaw, scaling, a10);
+}
+
+//----- (00063670) --------------------------------------------------------
+void sub_63670_draw_minimap_a(int16_t x, int16_t y, int16_t posX, int16_t posY, uint16_t width, uint16_t height, int16_t yaw, int16_t scaling, int a10)//244670
+{
+	int v10; // eax
+	int v11; // esi
+	int v12; // edi
+	int v13; // ecx
+	bool v14; // dl
+	__int16* v15; // ebx
+	int v16; // edx
+	__int16* v17; // esi
+	int v18; // ebx
+	__int16 v19; // ax
+	__int16* v20; // edx
+	__int16 v21; // ax
+	int8_t* v22x; // edx
+	int v23; // ebx
+	int v24; // esi
+	int v25; // eax
+	int v26; // eax
+	//int indexx; // eax
+	int v28; // ebx
+	int v29; // ecx
+	char* v30; // edi
+	int v31; // esi
+	int v32; // eax
+	int v33; // edx
+	char v34; // al
+	int v35; // ecx
+	int v36; // ebx
+	x_BYTE* v37; // edi
+	int v38; // ecx
+	int v39; // esi
+	int v40; // eax
+	int v41; // edx
+	int v42; // ebx
+	char* v43; // edi
+	int v44; // ecx
+	int v45; // esi
+	int v46; // eax
+	int v47; // edx
+	char v48; // al
+	int v49; // ecx
+	int v50; // ebx
+	x_BYTE* v51; // edi
+	int v52; // ecx
+	int v53; // esi
+	int v54; // eax
+	int v55; // edx
+	int v56; // ebx
+	char* v57; // edi
+	int v58; // ecx
+	int v59; // esi
+	int v60; // eax
+	int v61; // edx
+	char v62; // al
+	char* v63; // edi
+	uint8_t* v64x; // [esp+0h] [ebp-60h]
+	int v65; // [esp+4h] [ebp-5Ch]
+	int v66; // [esp+8h] [ebp-58h]
+	int v67; // [esp+Ch] [ebp-54h]
+	//int v68; // [esp+10h] [ebp-50h]
+	int v69; // [esp+14h] [ebp-4Ch]
+	uint8_t* v70x; // [esp+18h] [ebp-48h]
+	uint8_t* v71x; // [esp+1Ch] [ebp-44h]
+	int l; // [esp+20h] [ebp-40h]
+	int v73; // [esp+24h] [ebp-3Ch]
+	int i; // [esp+28h] [ebp-38h]
+	int k; // [esp+2Ch] [ebp-34h]
+	int m; // [esp+30h] [ebp-30h]
+	int j; // [esp+34h] [ebp-2Ch]
+	int v78; // [esp+38h] [ebp-28h]
+	__int16* v79; // [esp+3Ch] [ebp-24h]
+	int v80; // [esp+40h] [ebp-20h]
+	//uint8_t* v81z; // [esp+44h] [ebp-1Ch]
+	int v81y; // [esp+44h] [ebp-1Ch]
+	uint8_t* v82x; // [esp+48h] [ebp-18h]
+	int v83; // [esp+4Ch] [ebp-14h]
+	uint8_t* v84x; // [esp+50h] [ebp-10h]
+	int v85; // [esp+54h] [ebp-Ch]
+	__int16* v86; // [esp+58h] [ebp-8h]
+	bool v87; // [esp+5Ch] [ebp-4h]
+
+	v10 = x;
+	v11 = y;
+	v12 = width;
+	v13 = height;
+	v14 = !D41A0_0.m_GameSettings.str_0x2196.transparency_0x2198 && D41A0_0.terrain_2FECE.MapType != MapType_t::Cave;
+	v87 = v14;
+	if (x_WORD_180660_VGA_type_resolution & 1)
+	{
+		v10 = x >> 1;
+		v11 = y >> 1;
+		v12 = width >> 1;
+		v13 = height >> 1;
+		scaling *= 2;
+	}
+	v70x = &pdwScreenBuffer_351628[v11 * screenWidth_18062C + v10];
+	v71x = &x_DWORD_E9C3C[v11 * screenWidth_18062C + v10];
+	if (a10)
+	{
+		v15 = x_WORD_F4960;
+		v16 = v13;
+		while (v16)
+		{
+			v15 += 2;
+			*(v15 - 1) = 0;
+			v16--;
+			*(v15 - 2) = v12;
+		}
+	}
+	else
+	{
+		v78 = v13 / 2;
+		v17 = &x_WORD_F4960[2 * (v13 / 2)];
+		v79 = v17 - 2;
+		v73 = 0;
+		//v68 = v13 >> 1;
+		v18 = v12 >> 1;
+		v69 = 0x1000000 / (v13 >> 1);
+		while (v78)
+		{
+			v65 = v18 * (unsigned __int8)x_BYTE_F6EE0_tablesx[(0x14300 + v73) >> 16] >> 8;
+			v19 = v18 + v65;
+			v20 = v79;
+			*v17 = v18 + v65;
+			*v20 = v19;
+			v21 = v18 - v65;
+			v22x = (int8_t*)v79;
+			v17[1] = v18 - v65;
+			*(x_WORD*)(v22x + 2) = v21;
+			v79 = (__int16*)(v22x - 4);
+			v17 += 2;
+			v73 += v69;
+			v78--;
+		}
+	}
+	v23 = yaw & 0x7FF;
+	v24 = v13 * (scaling * Maths::sin_DB750[v23] >> 16);
+	v25 = scaling * (signed int)Maths::sin_DB750[0x200 + v23] >> 16;
+	v80 = scaling * Maths::sin_DB750[v23] >> 16;
+	v81y = v25;
+	v67 = v24 / v12;
+	v26 = v13 * v25;
+	v66 = v26 / v12;
+	v85 = posX - (v12 * (v26 / v12) - v24) / 2;
+	v86 = x_WORD_F4960;
+	v83 = posY - (v26 + v12 * (v24 / v12)) / 2;
+	v84x = v70x;
+	v82x = v71x;
+	if (isCaveLevel_D41B6)
+	{
+		if (a10)
+		{
+			for (i = v13; i; i--)
+			{
+				v33 = v86[1];
+				v28 = v85 + v33 * v66;
+				v29 = v83 + v33 * v67;
+				v30 = (char*)(v33 + v84x);
+				v64x = v33 + v82x;
+				v31 = *v86 - v86[1];
+				v32 = 0;
+				HIWORD(v33) = 0;
+				do
+				{
+					LOBYTE(v33) = BYTE1(v28);
+					BYTE1(v33) = BYTE1(v29);
+					v34 = 0;
+					if (!(mapAngle_13B4E0[v33] & 8))
+					{
+						LOBYTE(v32) = mapTerrainType_10B4E0[v33];
+						BYTE1(v33) = mapShading_12B4E0[v33];
+						LOBYTE(v33) = x_BYTE_F6EE0_tablesx[0x14000 + v32];
+						v34 = x_BYTE_F6EE0_tablesx[v33];
+					}
+					*v30++ = v34;
+					LOWORD(v28) = v66 + v28;
+					LOWORD(v29) = v67 + v29;
+					v31--;
+				} while (v31);
+				v84x += screenWidth_18062C;
+				v82x += screenWidth_18062C;
+				v86 += 2;
+				v85 -= v80;
+				//indexx = i - 1;
+				v83 += v81y;
+			}
+		}
+		else if (v87)
+		{
+			for (j = v13; j; j--)
+			{
+				v35 = v86[1];
+				v36 = v85 + v35 * v66;
+				v37 = (x_BYTE*)(v35 + v84x);
+				v64x = v35 + v82x;
+				v38 = v83 + v35 * v67;
+				v39 = *v86 - v86[1];
+				v40 = 0;
+				HIWORD(v41) = 0;
+				do
+				{
+					LOBYTE(v41) = BYTE1(v36);
+					BYTE1(v41) = BYTE1(v38);
+					if (!(mapAngle_13B4E0[v41] & 8))
+					{
+						LOBYTE(v40) = mapTerrainType_10B4E0[v41];
+						BYTE1(v41) = mapShading_12B4E0[v41];
+						LOBYTE(v41) = x_BYTE_F6EE0_tablesx[0x14000 + v40];
+						LOBYTE(v41) = x_BYTE_F6EE0_tablesx[v41];
+						BYTE1(v41) = *v37;
+						*v37 = x_BYTE_F6EE0_tablesx[0x4000 + v41];
+					}
+					v37++;
+					LOWORD(v36) = v66 + v36;
+					LOWORD(v38) = v67 + v38;
+					v39--;
+				} while (v39);
+				v84x += screenWidth_18062C;
+				v82x += screenWidth_18062C;
+				v86 += 2;
+				v85 -= v80;
+				//indexx = j - 1;
+				v83 += v81y;
+			}
+		}
+		else
+		{
+			for (k = v13; k; v83 += v81y)
+			{
+				v47 = v86[1];
+				v42 = v85 + v47 * v66;
+				v43 = (char*)(v47 + v84x);
+				v44 = v83 + v47 * v67;
+				v64x = v47 + v82x;
+				v45 = *v86 - v86[1];
+				v46 = 0;
+				HIWORD(v47) = 0;
+				do
+				{
+					LOBYTE(v47) = BYTE1(v42);
+					BYTE1(v47) = BYTE1(v44);
+					if (mapAngle_13B4E0[v47] & 8)
+					{
+						v48 = 0;
+					}
+					else
+					{
+						LOBYTE(v46) = mapTerrainType_10B4E0[v47];
+						BYTE1(v47) = mapShading_12B4E0[v47];
+						LOBYTE(v47) = x_BYTE_F6EE0_tablesx[0x14000 + v46];
+						v48 = x_BYTE_F6EE0_tablesx[v47];
+					}
+					*v43++ = v48;
+					LOWORD(v42) = v66 + v42;
+					LOWORD(v44) = v67 + v44;
+					v45--;
+				} while (v45);
+				v84x += screenWidth_18062C;
+				v82x += screenWidth_18062C;
+				v86 += 2;
+				//indexx = v81y;
+				v85 -= v80;
+				k--;
+			}
+		}
+	}
+	else if (a10)
+	{
+		for (l = v13; l; l--)
+		{
+			v49 = v86[1];
+			v50 = v85 + v49 * v66;
+			v51 = (x_BYTE*)(v49 + v84x);
+			v64x = v49 + v82x;
+			v52 = v83 + v49 * v67;
+			v53 = *v86 - v86[1];
+			v54 = 0;
+			HIWORD(v55) = 0;
+			do
+			{
+				LOBYTE(v55) = BYTE1(v50);
+				BYTE1(v55) = BYTE1(v52);
+				LOBYTE(v54) = mapTerrainType_10B4E0[v55];
+				BYTE1(v55) = mapShading_12B4E0[v55];
+				LOBYTE(v55) = x_BYTE_F6EE0_tablesx[0x14000 + v54];
+				*v51++ = x_BYTE_F6EE0_tablesx[v55];
+				v50 += v66;
+				v52 += v67;
+				--v53;
+			} while (v53);
+			v84x += screenWidth_18062C;
+			v82x += screenWidth_18062C;
+			v86 += 2;
+			v85 -= v80;
+			//indexx = l - 1;
+			v83 += v81y;
+		}
+	}
+	else
+	{
+		for (m = v13; m; v83 += v81y)
+		{
+			v61 = v86[1];
+			v56 = v85 + v61 * v66;
+			v57 = (char*)(v61 + v84x);
+			v58 = v83 + v61 * v67;
+			v64x = v61 + v82x;
+			v59 = *v86 - v86[1];
+			v60 = 0;
+			HIWORD(v61) = 0;
+			do
+			{
+				LOBYTE(v61) = BYTE1(v56);
+				BYTE1(v61) = BYTE1(v58);
+				LOBYTE(v60) = mapTerrainType_10B4E0[v61];
+				BYTE1(v61) = mapShading_12B4E0[v61];
+				LOBYTE(v61) = x_BYTE_F6EE0_tablesx[0x14000 + v60];
+				LOBYTE(v61) = x_BYTE_F6EE0_tablesx[v61];
+				BYTE1(v61) = *v57;
+				v62 = x_BYTE_F6EE0_tablesx[0x4000 + v61];
+				*v57 = v62;
+				v63 = x_InterlockedExchange((char**)&v64x, (v57 + 1));
+				*v63 = v62;
+				v57 = x_InterlockedExchange((char**)&v64x, (v63 + 1));
+				v56 += v66;
+				v58 += v67;
+				v59--;
+			} while (v59);
+			v84x += screenWidth_18062C;
+			v82x += screenWidth_18062C;
+			v86 += 2;
+			//indexx = v81y;
+			v85 -= v80;
+			m--;
+		}
+	}
+}
+
+//----- (00063C90) --------------------------------------------------------
+void sub_63C90_draw_minimap_b(int16_t x, int16_t y, int16_t posX, int16_t posY, uint16_t width, uint16_t height, int16_t yaw, int16_t scaling, int a10)//244c90
+{
+	//bool v14; // al
+	__int16* v15; // ebx
+	int v16; // edx
+	__int16* v17; // esi
+	int v18; // ebx
+	int v19; // ST00_4
+	int v20; // ebx
+	int v21; // esi
+	int v22; // eax
+	int v23; // eax
+	int v24; // edx
+	//int index; // eax
+	int v26; // ebx
+	int v27; // ecx
+	char* v28; // edi
+	int v29; // esi
+	int v30; // eax
+	int v31; // edx
+	char v32; // al
+	int v33; // ebx
+	int v34; // ecx
+	x_BYTE* v35; // edi
+	int v36; // esi
+	int v37; // eax
+	int v38; // edx
+	int v39; // ebx
+	int v40; // ecx
+	char* v41; // edi
+	int v42; // esi
+	int v43; // eax
+	int v44; // edx
+	char v45; // al
+	int v46; // ebx
+	int v47; // ecx
+	x_BYTE* v48; // edi
+	int v49; // esi
+	int v50; // eax
+	int v51; // edx
+	int v52; // ebx
+	int v53; // ecx
+	uint8_t* v54; // edi
+	int v55; // esi
+	int v56; // eax
+	int v57; // edx
+	int v58; // [esp+4h] [ebp-50h]
+	int v59; // [esp+8h] [ebp-4Ch]
+	uint8_t* ptrScreenBuffer; // [esp+14h] [ebp-40h]
+	int m; // [esp+18h] [ebp-3Ch]
+	int l; // [esp+1Ch] [ebp-38h]
+	int k; // [esp+20h] [ebp-34h]
+	int v64; // [esp+24h] [ebp-30h]
+	int i; // [esp+28h] [ebp-2Ch]
+	int j; // [esp+2Ch] [ebp-28h]
+	int v67; // [esp+30h] [ebp-24h]
+	__int16* v68; // [esp+34h] [ebp-20h]
+	int v69; // [esp+38h] [ebp-1Ch]
+	int v70; // [esp+3Ch] [ebp-18h]
+	int v72; // [esp+44h] [ebp-10h]
+	int v73; // [esp+48h] [ebp-Ch]
+	__int16* v74; // [esp+4Ch] [ebp-8h]
+	//bool v75; // [esp+50h] [ebp-4h]
+
+	//v14 = !x_D41A0_BYTEARRAY_0[8600] && x_D41A0_BYTEARRAY_0[196308] != 2;
+	//v75 = !x_D41A0_BYTEARRAY_0[8600] && x_D41A0_BYTEARRAY_0[196308] != 2;
+	if (x_WORD_180660_VGA_type_resolution & 1)
+	{
+		x = x >> 1;
+		y = y >> 1;
+		width = width >> 1;
+		height = height >> 1;
+		scaling *= 2;
+	}
+
+	ptrScreenBuffer = &pdwScreenBuffer_351628[screenWidth_18062C * y + x];
+	if (a10)
+	{
+		v15 = x_WORD_F4960;
+		v16 = height;
+		while (v16)
+		{
+			v15 += 2;
+			*(v15 - 1) = 0;
+			v16--;
+			*(v15 - 2) = width;
+		}
+	}
+	else
+	{
+		v67 = height / 2;//adress 244d1b
+		v17 = &x_WORD_F4960[2 * (height / 2)];
+		v68 = v17 - 2;
+		v64 = 0;
+		v18 = width >> 1;
+		while (v67)
+		{
+			v19 = v18 * x_BYTE_F6EE0_tablesx[0x14300 + (v64 >> 16)] >> 8;
+			v17[0] = v18 + v19;
+			v68[0] = v18 + v19;
+			v17[1] = v18 - v19;
+			v68[1] = v18 - v19;
+			v68 -= 2;
+			v17 += 2;
+			v64 += 0x1000000 / (height >> 1);
+			v67--;
+		}
+	}
+	v20 = yaw & 0x7FF;
+	v21 = height * (scaling * Maths::sin_DB750[v20] >> 16);
+	v22 = scaling * (signed int)Maths::sin_DB750[0x200 + v20] >> 16;
+	v70 = scaling * Maths::sin_DB750[v20] >> 16;
+	v69 = v22;
+	v59 = v21 / width;
+	v23 = height * v22;
+	v58 = v23 / width;
+	v72 = posX - (width * (v23 / width) - v21) / 2;
+	v74 = x_WORD_F4960;
+	v24 = posY - (v23 + width * (v21 / width)) / 2;
+	v73 = v24;
+	if (isCaveLevel_D41B6)//adress 244e53
+	{
+		if (a10)
+		{
+			for (i = height; i; v73 += v69)
+			{
+				v31 = v74[1];
+				v26 = v72 + v31 * v58;
+				v27 = v73 + v31 * v59;
+				v28 = (char*)(v31 + ptrScreenBuffer);
+				v29 = *v74 - v31;
+				v30 = 0;
+				HIWORD(v31) = 0;
+				do
+				{
+					LOBYTE(v31) = HIBYTE(v26);
+					HIBYTE(v31) = HIBYTE(v27);
+					v32 = 0;
+					if (!(mapAngle_13B4E0[v31] & 8))
+					{
+						LOBYTE(v30) = mapTerrainType_10B4E0[v31];
+						HIBYTE(v31) = mapShading_12B4E0[v31];
+						LOBYTE(v31) = x_BYTE_F6EE0_tablesx[0x14000 + v30];
+						v32 = x_BYTE_F6EE0_tablesx[v31];
+					}
+					*v28++ = v32;
+					LOWORD(v26) = v58 + v26;
+					LOWORD(v27) = v59 + v27;
+					v29--;
+				} while (v29);
+				ptrScreenBuffer += screenWidth_18062C;
+				v74 += 2;
+				//index = v69;
+				i--;
+				v72 -= v70;
+			}
+		}
+		else if (!D41A0_0.m_GameSettings.str_0x2196.transparency_0x2198 && D41A0_0.terrain_2FECE.MapType != MapType_t::Cave)
+		{
+			for (j = height; j; v73 += v69)
+			{
+				v38 = v74[1];
+				v33 = v72 + v38 * v58;
+				v34 = v73 + v38 * v59;
+				v35 = (x_BYTE*)(v38 + ptrScreenBuffer);
+				v36 = *v74 - v38;
+				v37 = 0;
+				HIWORD(v38) = 0;
+				do
+				{
+					LOBYTE(v38) = BYTE1(v33);
+					BYTE1(v38) = BYTE1(v34);
+					if (!(mapAngle_13B4E0[v38] & 8))
+					{
+						LOBYTE(v37) = mapTerrainType_10B4E0[v38];
+						BYTE1(v38) = mapShading_12B4E0[v38];
+						LOBYTE(v38) = x_BYTE_F6EE0_tablesx[0x14000 + v37];
+						LOBYTE(v38) = x_BYTE_F6EE0_tablesx[v38];
+						BYTE1(v38) = *v35;
+						*v35 = x_BYTE_F6EE0_tablesx[0x4000 + v38];
+					}
+					v35++;
+					LOWORD(v33) = v58 + v33;
+					LOWORD(v34) = v59 + v34;
+					v36--;
+				} while (v36);
+				ptrScreenBuffer += screenWidth_18062C;
+				v74 += 2;
+				//index = v69;
+				j--;
+				v72 -= v70;
+			}
+		}
+		else
+		{
+			for (k = height; k; v73 += v69)
+			{
+				v44 = v74[1];
+				v39 = v72 + v44 * v58;
+				v40 = v73 + v44 * v59;
+				v41 = (char*)(v44 + ptrScreenBuffer);
+				v42 = *v74 - v44;
+				v43 = 0;
+				HIWORD(v44) = 0;
+				do
+				{
+					LOBYTE(v44) = HIBYTE(v39);
+					HIBYTE(v44) = HIBYTE(v40);
+					if (mapAngle_13B4E0[v44] & 8)
+					{
+						v45 = 0;
+					}
+					else
+					{
+						LOBYTE(v43) = mapTerrainType_10B4E0[v44];
+						HIBYTE(v44) = mapShading_12B4E0[v44];
+						LOBYTE(v44) = x_BYTE_F6EE0_tablesx[0x14000 + v43];
+						v45 = x_BYTE_F6EE0_tablesx[v44];
+					}
+					*v41++ = v45;
+					LOWORD(v39) = v58 + v39;
+					LOWORD(v40) = v59 + v40;
+					v42--;
+				} while (v42);
+				ptrScreenBuffer += screenWidth_18062C;
+				v74 += 2;
+				//index = v69;
+				k--;
+				v72 -= v70;
+			}
+		}
+	}
+	else if (a10)
+	{
+		for (l = height; l; v73 += v69)
+		{
+			v51 = v74[1];
+			v46 = v72 + v51 * v58;
+			v47 = v73 + v51 * v59;
+			v48 = (x_BYTE*)(v51 + ptrScreenBuffer);
+			v49 = *v74 - v51;
+			v50 = 0;
+			HIWORD(v51) = 0;
+			do
+			{
+				LOBYTE(v51) = HIBYTE(v46);
+				HIBYTE(v51) = HIBYTE(v47);
+				LOBYTE(v50) = mapTerrainType_10B4E0[v51];
+				HIBYTE(v51) = mapShading_12B4E0[v51];
+				LOBYTE(v51) = x_BYTE_F6EE0_tablesx[0x14000 + v50];
+				*v48++ = x_BYTE_F6EE0_tablesx[v51];
+				v46 += v58;
+				v47 += v59;
+				v49--;
+			} while (v49);
+			ptrScreenBuffer += screenWidth_18062C;
+			v74 += 2;
+			//index = v69;
+			l--;
+			v72 -= v70;
+		}
+	}
+	else
+	{//adress 245156
+		for (m = height; m; v73 += v69)
+		{//0x2580=v71-pdwScreenBuffer_351628
+			v57 = v74[1];//adress 245185 mem edx 3ac764
+			v52 = v72 + v57 * v58;
+			v53 = v73 + v57 * v59;
+			v54 = &ptrScreenBuffer[v57];
+			v55 = v74[0] - v57;
+			v56 = 0;
+			HIWORD(v57) = 0;
+			do
+			{
+				LOBYTE(v57) = HIBYTE(v52);
+				HIBYTE(v57) = HIBYTE(v53);
+				LOBYTE(v56) = mapTerrainType_10B4E0[v57];
+				HIBYTE(v57) = mapShading_12B4E0[v57];
+				LOBYTE(v57) = x_BYTE_F6EE0_tablesx[0x14000 + v56];
+				LOBYTE(v57) = x_BYTE_F6EE0_tablesx[v57];
+				HIBYTE(v57) = v54[0];
+				*v54++ = x_BYTE_F6EE0_tablesx[0x4000 + v57];//here adress 2451b8
+				v52 += v58;
+				v53 += v59;
+				v55--;
+			} while (v55);
+			ptrScreenBuffer += screenWidth_18062C;
+			v74 += 2;
+			//index = v69;
+			m--;
+			v72 -= v70;
+		}
+	}
+}
+
+int debugcounter_644F0 = 0;
+
+
+//----- (00057FA0) --------------------------------------------------------
+void MovePlayer_57FA0_temporary_hack_for_gameui(axis_3d* position, unsigned __int16 a2, __int16 a3, __int16 a4)//238fa0 // move player
+{
+	// FIXME: this is a hack to make the game UI work
+	if (a4)
+	{
+		a3 &= 0x7ffu;
+		a2 &= 0x7ffu;
+		if (a3)
+		{
+			position->z -= (int)(a4 * Maths::sin_DB750[a3]) >> 16;//change z axis
+			a4 = (int)(a4 * Maths::sin_DB750[0x200 + a3]) >> 16;
+		}
+		position->x += (int)(a4 * Maths::sin_DB750[a2]) >> 16;
+		position->y -= (int)(a4 * Maths::sin_DB750[0x200 + a2]) >> 16;
+	}
+}
+
+//----- (000644F0) --------------------------------------------------------
+void DrawMinimapMarks_644F0(int16_t x, int16_t y, int16_t posX, int16_t posY, uint16_t width, uint16_t height, int16_t yaw, int16_t scaling)//2454f0
+{
+	signed int v8; // ebx
+	//__int64 v9; // rtt
+	int v9x;
+	int v10; // edi
+	int v11; // edx
+	int v12; // eax
+	__int16 v13; // ST10_2
+	char v14; // dl
+	type_event_0x6E8E* ix; // eax
+	axis_3d* v16x; // eax
+	axis_3d* v17x; // esi
+	int v18; // edx
+	signed int v19; // eax
+	signed int v20; // eax
+	int v21; // ebx
+	int v22; // ecx
+	int v23; // ebx
+	int v24; // ecx
+	int v25; // esi
+	signed int v26; // edx
+	int v27; // ecx
+	uint32_t v28; // ebx
+	int v29; // eax
+	unsigned __int8 v30; // cl
+	signed int v31; // ebx
+	int v32; // eax
+	int v33; // edi
+	//unsigned __int8 v34; // bl
+
+	/*
+	uint8_t origbyte20 = 0;
+	uint8_t remakebyte20 = 0;
+	int debugcounter11 = 0;
+	int comp20;
+
+	//if (debugcounter_644F0 >= 0x1c0)
+	{
+		comp20 = compare_with_sequence((char*)"002454F0-002DC4E0", (uint8_t*)x_BYTE_10B4E0_terraintype, 0x2dc4e0, debugcounter_644F0, 0x70000, 0x10000, &origbyte20, &remakebyte20);
+		comp20 = compare_with_sequence((char*)"002454F0-002DC4E0", (uint8_t*)x_BYTE_11B4E0_height, 0x2dc4e0, debugcounter_644F0, 0x70000, 0x10000, &origbyte20, &remakebyte20, 0x10000);
+		comp20 = compare_with_sequence((char*)"002454F0-002DC4E0", (uint8_t*)x_BYTE_12B4E0_shading, 0x2dc4e0, debugcounter_644F0, 0x70000, 0x10000, &origbyte20, &remakebyte20, 0x20000);
+		comp20 = compare_with_sequence((char*)"002454F0-002DC4E0", (uint8_t*)x_BYTE_13B4E0_angle, 0x2dc4e0, debugcounter_644F0, 0x70000, 0x10000, &origbyte20, &remakebyte20, 0x30000);
+		//comp20 = compare_with_sequence((char*)"00228320", (uint8_t*)x_BYTE_14B4E0, 0x2dc4e0, debugcounter11, 0x70000,0x10000, &origbyte20, &remakebyte20, 0x40000);
+		comp20 = compare_with_sequence((char*)"002454F0-002DC4E0", (uint8_t*)x_WORD_15B4E0_source, 0x2dc4e0, debugcounter_644F0, 0x70000, 0x20000, &origbyte20, &remakebyte20, 0x50000);
+
+		//uint8_t origbyte20 = 0;
+		//uint8_t remakebyte20 = 0;
+		comp20 = compare_with_sequence_D41A0((char*)"002454F0-00356038", (uint8_t*)& D41A0_BYTESTR_0, 0x356038, debugcounter_644F0, 224790, &origbyte20, &remakebyte20);
+
+		comp20 = compare_with_sequence_array_E2A74((char*)"002454F0-002B3A74", (uint8_t*)& array_E2A74, 0x2b3a74, debugcounter_644F0, 0xc4e, 0xc4e, &origbyte20, &remakebyte20);
+		comp20 = compare_with_sequence((char*)"002454F0-003514B0", (uint8_t*)& str_unk_1804B0ar, 0x3514b0, debugcounter_644F0, 0xab, 0xab, &origbyte20, &remakebyte20);
+	}
+
+	//if (debugcounter_644F0 >= 0x110)
+	{
+		debugcounter_644F0++;
+		debugcounter_644F0--;
+	}
+	comp20 = compare_with_sequence((char*)"002454F0-003AA0A4", pdwScreenBuffer_351628, 0x3aa0a4, debugcounter_47560, 320 * 200, 320 * 200, &origbyte20, &remakebyte20);
+	debugcounter_644F0++;
+	*/
+
+	axis_3d* v36x; // [esp-4h] [ebp-BEh]
+
+	int v37x[0xa8]; // [esp+0h] [ebp-BAh]
+	//int v38[5]; // [esp+4h] [ebp-B6h]//v37x[1]
+
+	//int v39; // [esp+18h] [ebp-A2h]//v37x[6]
+	//int v40; // [esp+1Ch] [ebp-9Eh]//v37x[7]
+	//int v41; // [esp+30h] [ebp-8Ah]//v37x[12]
+	//int v42; // [esp+34h] [ebp-86h]//v37x[13]
+	//int v43; // [esp+48h] [ebp-72h]//v37x[18]
+	//int v44; // [esp+4Ch] [ebp-6Eh]//v37x[19]
+	//int v45; // [esp+60h] [ebp-5Ah]//v37x[24]
+	//int v46; // [esp+64h] [ebp-56h]//v37x[25]
+	//int v47; // [esp+78h] [ebp-42h]//v37x[30]
+	//int v48; // [esp+7Ch] [ebp-3Eh]//v37x[31]
+	//int v49; // [esp+90h] [ebp-2Ah]//v37x[36]
+	//int v50; // [esp+94h] [ebp-26h]//v37x[37]
+
+	v51x_struct v51x; // [esp+A8h] [ebp-12h]//0
+	//__int16 v52; // [esp+ACh] [ebp-Eh]//+4
+	//uint8_t* v53; // [esp+AEh] [ebp-Ch]//+6
+	//int v54; // [esp+B2h] [ebp-8h]//+10
+	//int v55; // [esp+B6h] [ebp-4h]//+14
+	//int v56; // [esp+BAh] [ebp+0h]//+18
+	//int v57; // [esp+BEh] [ebp+4h]//+22
+	//int v58; // [esp+C2h] [ebp+8h]//+26
+	//int v59; // [esp+C6h] [ebp+Ch]//+30
+	//int v60; // [esp+CAh] [ebp+10h]//+34
+	//int v61; // [esp+CEh] [ebp+14h]//+38
+	//__int16 v62; // [esp+D2h] [ebp+18h]//+42
+	//__int16 v63; // [esp+D4h] [ebp+1Ah]//+44
+	//char v64; // [esp+D6h] [ebp+1Ch]//+46
+	axis_3d v65y; // [esp+D8h] [ebp+1Eh]
+	//__int16 v66; // [esp+DCh] [ebp+22h]
+	axis_3d v67x; // [esp+E0h] [ebp+26h]//+6
+	//__int16 v68; // [esp+E4h] [ebp+2Ah]//+10
+	int v69; // [esp+E8h] [ebp+2Eh]//+14
+	int v70; // [esp+ECh] [ebp+32h]//+18
+	int v71; // [esp+F0h] [ebp+36h]//+22
+	type_event_0x6E8E* v72x; // [esp+F4h] [ebp+3Ah]//+26
+	//int v73; // [esp+F8h] [ebp+3Eh]//+30
+	//int v74; // [esp+FCh] [ebp+42h]//34
+	type_event_0x6E8E* v75x; // [esp+100h] [ebp+46h]//38
+	int v76; // [esp+104h] [ebp+4Ah]//+40
+	int v77; // [esp+108h] [ebp+4Eh]
+	int v78; // [esp+10Ch] [ebp+52h]
+	int v79; // [esp+110h] [ebp+56h]
+	unsigned __int8 j; // [esp+114h] [ebp+5Ah]
+	char v81; // [esp+118h] [ebp+5Eh]
+	int v82; // [esp+148h] [ebp+8Eh]
+	int v83; // [esp+14Ch] [ebp+92h]
+
+	//fix it
+	v51x.v62xw_42 = 0;
+	v51x.v63xw_44 = 0;
+	v72x = 0;
+	//v66 = 0;
+	//fix it
+
+	v8 = 0x1000000;
+	v81 = 0;
+	if (x_WORD_180660_VGA_type_resolution & 1)
+	{
+		x >>= 1;
+		y >>= 1;
+		width >>= 1;
+		height >>= 1;
+		scaling *= 2;
+	}
+
+	v51x.v53xdw_6 = screenWidth_18062C * y + pdwScreenBuffer_351628 + x;
+	//LODWORD(v9) = 0x10000;
+	//HIDWORD(v9) = 0x10000 >> 31;
+	v9x = 0x10000;
+	v83 = v9x / scaling;
+	v51x.v56xdw_18 = posX;
+	v51x.v57xdw_22 = posY;
+	v51x.v58xdw_26 = width;
+	v51x.v59xdw_30 = height;
+	v51x.v60xdw_34 = width / 2;
+	v10 = yaw & 0x7FF;
+	v11 = v83 * Maths::sin_DB750[v10];
+	v51x.v61xdw_38 = height / 2;
+	v12 = v83 * Maths::sin_DB750[0x200 + v10] >> 16;
+	v51x.v55xdw_14 = -v11 >> 16;
+	v51x.v54xdw_10 = v12;
+	v82 = yaw & 0x7FF;
+	v13 = D41A0_0.LevelIndex_0xc;
+	v75x = x_DWORD_EA3E4[D41A0_0.array_0x2BDE[D41A0_0.LevelIndex_0xc].word_0x00a_2BE4_11240];
+	v14 = sub_595C0(v13);
+	switch (v14 + 3)
+	{
+	case 0:
+	case 1:
+	case 2:
+		if (v14 == -3)
+		{
+			for (ix = x_D41A0_BYTEARRAY_4_struct.dword_38535; ; ix = ix->next_0)
+			{
+				v72x = ix;
+				if (ix <= x_DWORD_EA3E4[0])
+					break;
+				//LOBYTE(i) = (uint8_t)v72;//can remove it
+				if (ix->model_0x40_64 == 31)
+				{
+					v81 = 1;
+					break;
+				}
+			}
+		}
+		else
+		{
+			for (ix = x_D41A0_BYTEARRAY_4_struct.dword_38535; ; ix = ix->next_0)
+			{
+				v72x = ix;
+				if (ix <= x_DWORD_EA3E4[0])
+					break;
+				//LOBYTE(i) = (uint8_t)v72;//can remove it
+				if (ix->model_0x40_64 == 12)
+				{
+					v81 = 1;
+					break;
+				}
+			}
+		}
+		if (v81)
+		{
+			v36x = &v67x;
+			v16x = &v75x->axis_0x4C_76;
+			v17x = &v72x->axis_0x4C_76;
+			goto LABEL_47;
+		}
+		break;
+	case 4:
+	case 5:
+		//LOBYTE(i) = (uint8_t)sub_59610(v72x, D41A0_BYTESTR_0.word_0xc);
+		if (sub_59610(&v72x, D41A0_0.LevelIndex_0xc))
+		{
+			v51x.v64xb_46 = (*xadataclrd0dat.colorPalette_var28)[3840];
+			v51x.v51y = v72x->axis_0x4C_76;
+			if (sub_64CE0_draw_follow_rectangle(&v51x))
+				sub_885E0(v72x, v51x.v62xw_42, v51x.v63xw_44, 0x52u);
+			v36x = &v72x->axis_0x4C_76;
+			v17x = &v72x->axis_0x4C_76;
+			v16x = &v75x->axis_0x4C_76;
+			v81 = 1;
+			goto LABEL_47;
+		}
+		break;
+	case 8:
+		//LOBYTE(i) = sub_596C0(v65x, D41A0_BYTESTR_0.word_0xc);
+		if (sub_596C0(&v65y, D41A0_0.LevelIndex_0xc))
+		{
+			v51x.v64xb_46 = (*xadataclrd0dat.colorPalette_var28)[0xff0];
+			v51x.v51y = v65y;
+			if (sub_64CE0_draw_follow_rectangle(&v51x))
+				sub_885E0(0, v51x.v62xw_42, v51x.v63xw_44, 0x51u);
+			v36x = &v65y;
+			v17x = &v65y;
+			v16x = &v75x->axis_0x4C_76;
+			v81 = 1;
+		LABEL_47:
+			v67x = *v17x;
+			//i = (uint8_t*)Maths::sub_58490_radix_3d_2(v16x, v36x);
+			v8 = Maths::sub_58490_radix_3d_2(v16x, v36x);
+		}
+		break;
+	case 0xA:
+		v18 = D41A0_0.stages_0x3654C[D41A0_0.struct_0x3659C[D41A0_0.LevelIndex_0xc].substr_3659C.ObjectiveText_1].str_36552_un.dword;
+		v51x.v64xb_46 = (*xadataclrd0dat.colorPalette_var28)[3840];
+		for (ix = x_D41A0_BYTEARRAY_4_struct.bytearray_38403x[v18]; ; ix = ix->next_0)
+		{
+			v72x = ix;
+			if (ix <= x_DWORD_EA3E4[0])
+				break;
+			v51x.v51y = v72x->axis_0x4C_76;
+			if (sub_64CE0_draw_follow_rectangle(&v51x))
+				sub_885E0(v72x, v51x.v62xw_42, v51x.v63xw_44, 0x52u);
+			v19 = Maths::sub_58490_radix_3d_2(&v75x->axis_0x4C_76, &v72x->axis_0x4C_76);
+			if (v19 < v8)
+			{
+				v81 = 1;
+				v8 = v19;
+				v67x = v72x->axis_0x4C_76;
+			}
+		}
+		break;
+	case 0xC:
+		v51x.v64xb_46 = (*xadataclrd0dat.colorPalette_var28)[3840];
+		v76 = D41A0_0.stages_0x3654C[D41A0_0.struct_0x3659C[D41A0_0.LevelIndex_0xc].substr_3659C.ObjectiveText_1].str_36552_un.dword;
+		for (ix = x_D41A0_BYTEARRAY_4_struct.dword_38527; ; ix = ix->next_0)
+		{
+			v72x = ix;
+			if (ix <= x_DWORD_EA3E4[0])
+				break;
+			v79 = v76;
+			for (j = 0; v79 && j < 8u; j++)
+			{
+				if (v72x->byte_0x46_70 == v79)
+				{
+					v51x.v51y = v72x->axis_0x4C_76;
+					if (sub_64CE0_draw_follow_rectangle(&v51x))
+						sub_885E0(v72x, v51x.v62xw_42, v51x.v63xw_44, 0x52u);
+					v20 = Maths::sub_58490_radix_3d_2(&v75x->axis_0x4C_76, &v72x->axis_0x4C_76);
+					if (v20 < v8)
+					{
+						v8 = v20;
+						v81 = 1;
+						v67x = v72x->axis_0x4C_76;
+					}
+				}
+				v79 = str_D93C0_bldgprmbuffer[v79].byte_3;
+			}
+		}
+		break;
+	default:
+		break;
+	}
+	if (v81)
+	{
+		//LOBYTE(i) = (uint8)x_D41A0_BYTEARRAY_4;
+		if (x_D41A0_BYTEARRAY_4_struct.byteindex_26 & 0x40)
+		{
+			if (x_D41A0_BYTEARRAY_4_struct.byteindex_121[6])
+			{
+				v21 = v8 - 512;
+				LOWORD(v78) = Maths::sub_581E0_maybe_tan2(&v75x->axis_0x4C_76, &v67x);
+				v78 = (unsigned __int16)v78;
+				if (v21 > 15872)
+					LOWORD(v21) = 15872;
+				v67x = v75x->axis_0x4C_76;
+				MovePlayer_57FA0_temporary_hack_for_gameui(&v67x, v78, 0, v21);
+				v22 = v83 * Maths::sin_DB750[v82];
+				v23 = v83 * (x_DWORD)Maths::sin_DB750[0x200 + v82];
+				v71 = (signed __int16)(*(int16_t*)&v67x - posX);
+				v23 >>= 16;
+				v24 = -v22 >> 16;
+				v25 = width / 2 + ((v23 * v71 - v24 * (signed __int16)((int16_t)v67x.y - posY)) >> 16);
+				v37x[7] = 13;
+				v37x[13] = 13;
+				v37x[19] = 13;
+				v37x[37] = 13;
+				v37x[6] = 9;
+				v77 = height / 2 + ((v71 * v24 + (signed __int16)((int16_t)v67x.y - posY) * v23) >> 16);
+				v37x[0] = 0;
+				v37x[1] = 0;
+				v37x[12] = -9;
+				v37x[30] = -4;
+				v37x[36] = -4;
+				v37x[25] = 23;
+				v37x[31] = 23;
+				v37x[18] = 4;
+				v37x[24] = 4;
+				v26 = 1;
+				v78 = ((x_WORD)v78 - (x_WORD)v82) & 0x7FF;
+				v27 = Maths::sin_DB750[v78];
+				v28 = Maths::sin_DB750[0x200 + v78];
+				while (v26 < 7)
+				{
+					v29 = 3 * v26;
+					v70 = 16 * v37x[6 * v26];
+					v69 = 16 * v37x[1 + 6 * v26];
+					v37x[2 * v29] = ((signed int)v28 * v70 - v27 * v69) >> 20;
+					//v73 = v27 * v70;
+					//v74 = (x_DWORD)v28 * v69;
+					v26++;
+					v37x[1 + 2 * v29] = ((signed int)v28 * v69 + v27 * v70) >> 20;
+				}
+				if (x_WORD_180660_VGA_type_resolution & 1)
+				{
+					v25 *= 2;
+					v77 *= 2;
+				}
+				v30 = (*xadataclrd0dat.colorPalette_var28)[3840];
+				v31 = 0;
+				while (v31 < 7)
+				{
+					v32 = 3 * v31;
+					v37x[2 * v32] = v25 + x + v37x[6 * v31];
+					v33 = v37x[1 + (6 * v31++)];
+					v37x[1 + 2 * v32] = v77 + y + v33;
+				}
+				//adress 245be1
+				//v34 = v30;
+				sub_2BD10_draw_line(v37x[0], v37x[1], v37x[6], v37x[7], v30);
+				sub_2BD10_draw_line(v37x[6], v37x[7], v37x[18], v37x[19], v30);
+				sub_2BD10_draw_line(v37x[18], v37x[19], v37x[24], v37x[25], v30);
+				sub_2BD10_draw_line(v37x[24], v37x[25], v37x[30], v37x[31], v30);
+				sub_2BD10_draw_line(v37x[30], v37x[31], v37x[36], v37x[37], v30);
+				sub_2BD10_draw_line(v37x[36], v37x[37], v37x[12], v37x[13], v30);
+				sub_2BD10_draw_line(v37x[12], v37x[13], v37x[0], v37x[1], v30);
+			}
+		}
+	}
+}
+
+//----- (00064CE0) --------------------------------------------------------
+char sub_64CE0_draw_follow_rectangle(v51x_struct* a1)//245ce0
+{
+	int16_t v1; // ecx
+	int16_t v2; // esi
+	char v3; // ah
+	int v4; // ebx
+	char v5; // al
+	int v6; // ecx
+	int v7; // esi
+	int v8; // ecx
+	int v9; // ebx
+	x_BYTE* v10; // edx
+	x_BYTE* v11; // edx
+	x_BYTE* v12; // edx
+	x_BYTE* v13; // edx
+	x_BYTE* v14; // edx
+	x_BYTE* v15; // edx
+
+	v1 = a1->v51y.x - a1->v56xdw_18;
+	v2 = a1->v51y.y - a1->v57xdw_22;
+	v3 = 0;
+	v4 = a1->v60xdw_34 + ((int32_t)(v1 * a1->v54xdw_10 - v2 * a1->v55xdw_14) >> 16);
+	v5 = a1->v64xb_46;
+	if (v4 >= 0 && v4 < a1->v58xdw_26)
+	{
+		v6 = a1->v61xdw_38 + (int16_t)((a1->v54xdw_10 * v2 + a1->v55xdw_14 * v1) >> 16);
+		if (v6 >= 3 && v6 < a1->v59xdw_30 - 3 && v4 >= x_WORD_F4960[1 + 2 * v6] + 3 && v4 < x_WORD_F4960[2 * v6] - 3)
+		{
+			a1->v62xw_42 = v4;
+			a1->v63xw_44 = v6;
+			v3 = 1;
+			if (!(x_D41A0_BYTEARRAY_4_struct.byteindex_26 & 3))
+			{
+				v7 = screenWidth_18062C;
+				v8 = screenWidth_18062C * (v6 - 1);
+				v9 = v4 - 1;
+				v10 = (x_BYTE*)(v8 + v9 + pdwScreenBuffer_351628);
+				v10[0] = v5;
+				v10[1] = v5;
+				v10[2] = v5;
+				v11 = &v10[v7];
+				v11[0] = v5;
+				v11[2] = v5;
+				v12 = &v11[v7];
+				v12[0] = v5;
+				v12[1] = v5;
+				v12[2] = v5;
+				if (1 == D41A0_0.m_GameSettings.m_Display.m_uiScreenSize)
+				{
+					v13 = (x_BYTE*)(v8 + v9 + x_DWORD_E9C3C);
+					v13[0] = v5;
+					v13[1] = v5;
+					v13[2] = v5;
+					v14 = &v13[v7];
+					v14[0] = v5;
+					v14[2] = v5;
+					v15 = &v14[v7];
+					v15[0] = v5;
+					v15[1] = v5;
+					v15[2] = v5;
+				}
+			}
+		}
+	}
+	return v3;
+}
+
+int debugcounter_2695e0 = 0;
+
+//----- (000885E0) --------------------------------------------------------
+void sub_885E0(type_event_0x6E8E* a1x, int16_t posX, int16_t posY, uint16_t a4)//2695e0
+{
+	signed int v4; // eax
+	int v5; // edx
+	//unsigned int v6; // ebx
+	//int v7; // ebx
+	//char v8; // al
+	//char v9; // dl
+	//char v10; // dh
+	char v12; // [esp+0h] [ebp-4h]
+
+	if (debugcounter_2695e0 == 0xa)
+	{
+		debugcounter_2695e0++;
+		debugcounter_2695e0--;
+	}
+
+	//LOBYTE(v4) = (uint8)x_D41A0_BYTEARRAY_4;
+	v5 = 0;
+	if (x_D41A0_BYTEARRAY_4_struct.showHelp_10)
+	{
+		//LOBYTE(v4) = a4;
+		if (!(str_E2A74[a4].axis_2[0] & 8))
+		{
+			if (x_WORD_180660_VGA_type_resolution & 1)
+			{
+				posX *= 2;
+				posY *= 2;
+			}
+			//LOBYTE(v4) = a4;
+			//v6 = 30 * (signed __int16)a4;
+			if (str_E2A74[a4].axis_2[0] & 2)
+			{
+				if (a1x == str_E2A74[a4].dword_12 && !(*(int16_t*)&str_unk_1804B0ar.PopupStatusByte_0x9e & 0x20))
+				{
+					//v9 = array_E2A74[0x2 + 30 * a4] | 0x10;
+					str_E2A74[a4].axis_2[3] = posX;
+					str_E2A74[a4].axis_2[0] |= 0x10;
+					//LOBYTE(v4) = a3;
+					//v10 = HIBYTE(*(int16_t*)&str_unk_1804B0ar.PopupStatusByte_0x9e) | 2;
+					str_E2A74[a4].axis_2[4] = posY;
+					str_unk_1804B0ar.byte_0x9f |= 2;
+				}
+			}
+			else
+			{
+				v12 = 0;
+				if (!a1x
+					|| (v4 = Maths::sub_58490_radix_3d_2(
+						&x_DWORD_EA3E4[D41A0_0.array_0x2BDE[D41A0_0.LevelIndex_0xc].word_0x00a_2BE4_11240]->axis_0x4C_76,
+						&a1x->axis_0x4C_76),
+						v5 = v4,
+						!str_E2A74[a4].dword_12)
+					|| v4 < str_E2A74[a4].dword_20)
+				{
+					v12 = 1;
+				}
+				if (v12)
+				{
+					//v7 = 15 * (signed __int16)a4;
+					str_E2A74[a4].dword_12 = a1x;//xx
+					//v8 = array_E2A74[0x2 + 30 * a4];
+					str_E2A74[a4].dword_20 = v5;
+					//LOBYTE(v4) = v8 | 0x10;
+					str_E2A74[a4].axis_2[0] |= 0x10;
+				}
+			}
+		}
+		if (a4 >= 0x51u && a4 <= 0x53u)
+		{
+			//LOBYTE(v4) = a4;
+			str_E2A74[a4].axis_2[0] |= 0x40u;
+		}
+	}
+
+	debugcounter_2695e0++;
+}
+
+//----- (00089420) --------------------------------------------------------
+void ComputeTextboxSizesFromTextWords_89420(Type_TextBox_1804B0* textbox, const char* text, uint8_t scale)//26a420
+{
+	int lineCharWidth; // eax
+	unsigned int textBufferLenght; // kr04_4
+	signed int textWidth; // ecx
+	signed int textboxWidth; // ebx
+	int lastLineIndex;
+	int textIndex;
+	signed __int16 countOfLines; // [esp+4h] [ebp-4h]
+
+	textBufferLenght = strlen(text) + 1;
+	countOfLines = 1;
+	textbox->textBoxWidth_0x4 = 0;
+	textbox->textboxHeight_0x6 = 0;
+	if (textBufferLenght != 1)
+	{
+		textWidth = (textbox->charWidth_0x10 * scale) * strlen(text);
+		textboxWidth = textbox->maxTextboxWidth2_0xc;
+		if (textWidth <= textboxWidth)
+		{
+			textbox->textBoxWidth_0x4 = textWidth;
+		}
+		else
+		{
+			if (textWidth % textboxWidth < textboxWidth / 2)//last line is lower than halfscreen
+			{
+				while (1)
+				{
+					if (textWidth % textboxWidth >= textboxWidth / 2 || textboxWidth <= 0)
+						break;
+					textboxWidth -= textbox->charWidth_0x10 * scale;
+				}
+			}
+			if (textboxWidth > 0)
+			{
+				lineCharWidth = (textboxWidth + (textbox->charWidth_0x10 * scale) - 1) / (textbox->charWidth_0x10 * scale);
+				countOfLines = 0;
+				lastLineIndex = 0;
+				while (lastLineIndex < textBufferLenght - 1)
+				{
+					for (textIndex = lastLineIndex + lineCharWidth; text[lastLineIndex + textIndex] != 32 && textIndex < textBufferLenght - 1; textIndex--);//to previous word
+					lastLineIndex = textIndex + 1;
+					countOfLines++;
+				}
+				textbox->textBoxWidth_0x4 = textboxWidth;
+			}
+		}
+		textbox->textboxHeight_0x6 = (textbox->charHeight_0x12 * scale) * countOfLines;
+	}
+}
+
+//----- (00089520) --------------------------------------------------------
+void ConstrainTextboxSizes_89520(Type_TextBox_1804B0* textbox, uint8_t scale)//26a520
+{
+	int textboxHeight; // edi
+	int centerToMaxWidth; // ecx
+	int centerToMaxHeight; // esi
+	int textBoxWidth; // [esp+0h] [ebp-Ch]
+
+	int frameCornerWidth = (*filearray_2aa18c[filearrayindex_MSPRD00DATTAB].posistruct)[FRAME_CORNER].width_4 * scale;
+	int frameCornerHeight = (*filearray_2aa18c[filearrayindex_MSPRD00DATTAB].posistruct)[FRAME_CORNER].height_5 * scale;
+
+	textBoxWidth = textbox->textBoxWidth_0x4;
+	textboxHeight = textbox->textboxHeight_0x6;
+	centerToMaxWidth = textbox->maxTextboxWidth_0x0 - (textBoxWidth / 2);
+	centerToMaxHeight = textbox->maxTextboxHeight_0x2 - (textboxHeight / 2);
+
+	if (centerToMaxWidth < textbox->minPosX_0x14 + frameCornerWidth)
+		centerToMaxWidth = textbox->minPosX_0x14 + frameCornerWidth;
+
+	if (centerToMaxHeight < textbox->minPosY_0x18 + frameCornerHeight)
+		centerToMaxHeight = textbox->minPosY_0x18 + frameCornerHeight;
+
+	if (centerToMaxWidth + textBoxWidth >= textbox->maxPosX_0x16)
+		centerToMaxWidth = textbox->maxPosX_0x16 - textBoxWidth;
+
+	if (centerToMaxHeight + textboxHeight >= textbox->maxPosY_0x1a)
+		centerToMaxHeight = textbox->maxPosY_0x1a - textboxHeight;
+
+	textbox->textboxPosX_0x8 = centerToMaxWidth;
+	textbox->textboxPosY_0xa = centerToMaxHeight;
+	textbox->maxTextboxWidth_0x0 = centerToMaxWidth + (textBoxWidth / 2);
+	textbox->maxTextboxHeight_0x2 = centerToMaxHeight + (textboxHeight / 2);
+}
+
+
+//----- (000895D0) --------------------------------------------------------
+void DrawTextboxText_895D0(Type_TextBox_1804B0* textbox, const char* text, uint8_t scale)//26a5d0
+{
+	int i;
+	unsigned __int8 color;
+	int textboxPosX;
+	char buffer[1000]; // [esp+0h] [ebp-12h]
+	memset(buffer, NULL, 1000 * sizeof(*buffer));
+	int lastPos = 0;
+	int posY = textbox->textboxPosY_0xa;
+	int textlen = strlen(text);
+	int width = textbox->textBoxWidth_0x4 / (textbox->charWidth_0x10 * scale);
+	if (textlen > 0)
+	{
+		do
+		{
+			for (i = width + lastPos; text[i] != 32 && i < textlen; i--);
+			memcpy(buffer, &text[lastPos], i - lastPos);
+			buffer[i] = 0;
+			color = textbox->color1_0x30;
+			textboxPosX = textbox->textboxPosX_0x8;
+			lastPos = i + 1;
+			DrawText_2BC10(buffer, textboxPosX, posY, color, scale);
+			posY += textbox->charHeight_0x12 * scale;
+		} while (textlen > lastPos);
+	}
+}
+
+
+//----- (00089690) --------------------------------------------------------
+void DrawTextboxFrame_89690(Type_TextBox_1804B0* textbox, uint8_t scale)//26a690
+{
+	int width1 = (*filearray_2aa18c[filearrayindex_MSPRD00DATTAB].posistruct)[FRAME_TOP_BOTTOM_SIDE].width_4 * scale;
+	int width2 = (*filearray_2aa18c[filearrayindex_MSPRD00DATTAB].posistruct)[FRAME_LEFT_RIGHT_SIDE].width_4 * scale;
+	int height1 = (*filearray_2aa18c[filearrayindex_MSPRD00DATTAB].posistruct)[FRAME_CORNER].height_5 * scale;
+	int height2 = (*filearray_2aa18c[filearrayindex_MSPRD00DATTAB].posistruct)[FRAME_LEFT_RIGHT_SIDE].height_5 * scale;
+	int beginX = textbox->framePosX_0x24 - width2;
+	int endX = textbox->frameWidth_0x28 + textbox->framePosX_0x24;
+	int beginYPlus = textbox->framePosY_0x26 - ((*filearray_2aa18c[filearrayindex_MSPRD00DATTAB].posistruct)[FRAME_TOP_BOTTOM_SIDE].height_5 * scale);
+	int endY = textbox->frameHeight_0x2a + textbox->framePosY_0x26;
+	int countYchars = textbox->textboxHeight_0x6 / (textbox->charHeight_0x12 * scale);
+	//Top Left/Right corners
+	sub_2BB40_draw_bitmap(textbox->framePosX_0x24 - width2, beginYPlus, (*filearray_2aa18c[filearrayindex_MSPRD00DATTAB].posistruct)[FRAME_CORNER], scale);
+	sub_2BB40_draw_bitmap(endX, beginYPlus, (*filearray_2aa18c[filearrayindex_MSPRD00DATTAB].posistruct)[FRAME_CORNER], scale);
+	int indexY = height1 + beginYPlus;
+
+	//Sides Left/Right 
+	for (int i = 0; i < countYchars; i++)
+	{
+		sub_2BB40_draw_bitmap(beginX, indexY, (*filearray_2aa18c[filearrayindex_MSPRD00DATTAB].posistruct)[FRAME_LEFT_RIGHT_SIDE], scale);
+		sub_2BB40_draw_bitmap(endX, indexY, (*filearray_2aa18c[filearrayindex_MSPRD00DATTAB].posistruct)[FRAME_LEFT_RIGHT_SIDE], scale);
+		indexY += height2;
+	}
+
+	//Bottom Left/Right corners
+	sub_2BB40_draw_bitmap(beginX, indexY, (*filearray_2aa18c[filearrayindex_MSPRD00DATTAB].posistruct)[FRAME_CORNER], scale);
+	sub_2BB40_draw_bitmap(endX, indexY, (*filearray_2aa18c[filearrayindex_MSPRD00DATTAB].posistruct)[FRAME_CORNER], scale);
+
+	//Sides Top/Bottom 
+	for (int i = beginX + width2; i < endX; i += width1)
+	{
+		sub_2BB40_draw_bitmap(i, beginYPlus, (*filearray_2aa18c[filearrayindex_MSPRD00DATTAB].posistruct)[FRAME_TOP_BOTTOM_SIDE], scale);
+		sub_2BB40_draw_bitmap(i, endY, (*filearray_2aa18c[filearrayindex_MSPRD00DATTAB].posistruct)[FRAME_TOP_BOTTOM_SIDE], scale);
+	}
+}
+
+
+//----- (00089980) --------------------------------------------------------
+void ComputeFrameSizes_89980(Type_TextBox_1804B0* textbox, uint8_t scale)//26a980
+{
+	int width = (*filearray_2aa18c[filearrayindex_MSPRD00DATTAB].posistruct)[FRAME_TOP_BOTTOM_SIDE].width_4 * scale;
+
+	int height1 = (*filearray_2aa18c[filearrayindex_MSPRD00DATTAB].posistruct)[FRAME_TOP_BOTTOM_SIDE].height_5 * scale;
+	int height2 = (*filearray_2aa18c[filearrayindex_MSPRD00DATTAB].posistruct)[FRAME_LEFT_RIGHT_SIDE].height_5 * scale;
+	int height3 = (*filearray_2aa18c[filearrayindex_MSPRD00DATTAB].posistruct)[FRAME_CORNER].height_5 * scale;
+	textbox->frameWidth_0x28 = width * (((width) - (1 * scale) + textbox->textBoxWidth_0x4 + (8 * scale)) / (width));
+	textbox->framePosX_0x24 = textbox->textboxPosX_0x8 - (textbox->frameWidth_0x28 - textbox->textBoxWidth_0x4) / 2;
+	textbox->frameHeight_0x2a = (height2 - (1 * scale) + textbox->textboxHeight_0x6) / height2 * height2 + 2 * (height3 - height1);
+	textbox->framePosY_0x26 = textbox->textboxPosY_0xa - (textbox->frameHeight_0x2a - textbox->textboxHeight_0x6) / 2;
+}
+
+//----- (000BD524) --------------------------------------------------------
+void sub_BD524(int16_t a1, int16_t a2, int16_t a3, int16_t a4, char a5)//29e524
+{
+	sub_BD542(a1 / 2, a2 / 2, a3 / 2, a4 / 2, a5);
+}
+
+//----- (000BD542) --------------------------------------------------------
+void sub_BD542(uint16_t a1, uint16_t a2, uint16_t a3, uint16_t a4, char a5)//29e542
+{
+	__int16 v5; // bx
+	__int16 v6; // cx
+	unsigned __int16 v7; // cx
+	unsigned __int16 v8; // ax
+	unsigned __int16 v9; // ax
+	_BYTE* v10; // edi
+	int v11; // esi
+	unsigned __int16 v12; // ax
+	unsigned __int16 v13; // cx
+	unsigned __int16 v14; // si
+	__int16 v15; // dx
+	__int16 v16; // bx
+	__int16 v17; // cx
+	int v18; // edi
+	unsigned int v19; // ecx
+	__int16 v20; // ax
+	int v21; // eax
+	int v22; // ecx
+	__int16 v23; // cx
+	_BYTE* v24; // edi
+	int v25; // [esp+0h] [ebp-Ch]
+	int v26; // [esp+4h] [ebp-8h]
+	uint8_t* v27; // [esp+8h] [ebp-4h]
+
+	v27 = &pdwScreenBuffer_351628[x_DWORD_18063C_sprite_sizex + screenWidth_18062C * x_DWORD_180650_positiony];
+	v5 = a2;
+	if ((a2 & 0x8000u) == 0)
+	{
+		v7 = a4;
+		if ((a4 & 0x8000u) == 0)
+		{
+			if (a2 == a4)
+			{
+				if ((__int16)a2 >= (__int16)x_DWORD_180644_map_resolution2_y)
+					return;
+				v18 = a1;
+				v19 = a3;
+				if (a3 < a1)
+				{
+					LOWORD(v19) = a1;
+					LOWORD(v18) = a3;
+				}
+				v20 = x_DWORD_180648_map_resolution2_x - 1;
+				if ((v18 & 0x8000u) == 0)
+				{
+					if ((__int16)v18 > v20)
+						return;
+					if ((unsigned __int16)v19 > (unsigned __int16)v20)
+						LOWORD(v19) = x_DWORD_180648_map_resolution2_x - 1;
+				}
+				else
+				{
+					if ((__int16)v19 <= 0)
+						return;
+					LOWORD(v18) = 0;
+					if ((unsigned __int16)v19 > (unsigned __int16)v20)
+						LOWORD(v19) = x_DWORD_180648_map_resolution2_x - 1;
+				}
+				LOWORD(v19) = v19 - v18;
+				goto LABEL_54;
+			}
+			if ((__int16)(a2 - a4) < 0)
+			{
+				if ((__int16)a2 >= (__int16)x_DWORD_180644_map_resolution2_y)
+					return;
+				if ((__int16)a4 >= (__int16)x_DWORD_180644_map_resolution2_y)
+					goto LABEL_5;
+			}
+			else
+			{
+				if ((__int16)a4 >= (__int16)x_DWORD_180644_map_resolution2_y)
+					return;
+				a4 = a2;
+				a2 = v7;
+				v9 = a1;
+				a1 = a3;
+				a3 = v9;
+				if (v5 >= (__int16)x_DWORD_180644_map_resolution2_y)
+					goto LABEL_5;
+			}
+			goto LABEL_15;
+		}
+		a4 = a2;
+		v5 = v7;
+		v6 = a2;
+		v8 = a1;
+		a1 = a3;
+		a3 = v8;
+	}
+	else
+	{
+		v6 = a4;
+	}
+	if (v6 < 0)
+		return;
+	a1 += (__int16)-v5 * (__int16)(a3 - a1) / (__int16)(v6 - v5);
+	a2 = 0;
+	if ((__int16)(a4 - x_DWORD_180644_map_resolution2_y) >= 0)
+	{
+	LABEL_5:
+		a3 = a1 + (__int16)(x_DWORD_180644_map_resolution2_y - a2) * (__int16)(a3 - a1) / (__int16)(a4 - a2);
+		a4 = x_DWORD_180644_map_resolution2_y - 1;
+	}
+LABEL_15:
+	if ((a1 & 0x8000u) == 0)
+	{
+		if ((__int16)(a1 - x_DWORD_180648_map_resolution2_x) < 0)
+		{
+			if ((a3 & 0x8000u) == 0)
+			{
+				if ((__int16)(a3 - x_DWORD_180648_map_resolution2_x) < 0)
+				{
+					v26 = 1;
+					if (a1 == a3)
+					{
+						v21 = a2;
+						v22 = a4;
+						if (a4 >= a2)
+						{
+							v23 = a4 - a2;
+						}
+						else
+						{
+							LOWORD(v21) = a4;
+							v23 = a2 - a4;
+						}
+						v24 = (_BYTE*)(v27 + screenWidth_18062C * v21 + a1);
+						LOWORD(v22) = v23 + 1;
+						do
+						{
+							*v24 = a5;
+							v24 += screenWidth_18062C;
+							--v22;
+						} while (v22);
+						return;
+					}
+					if ((__int16)(a1 - a3) >= 0)
+						v26 = -1;
+				}
+				else
+				{
+					v26 = 1;
+					a4 = a2 + (__int16)(x_DWORD_180648_map_resolution2_x - a1) * (__int16)(a4 - a2) / (__int16)(a3 - a1);
+					a3 = x_DWORD_180648_map_resolution2_x - 1;
+				}
+			}
+			else
+			{
+				v26 = -1;
+				a4 = a2 + (__int16)a1 * (__int16)(a4 - a2) / (__int16)(a1 - a3);
+				a3 = 0;
+			}
+		}
+		else
+		{
+			if ((__int16)a3 >= (__int16)x_DWORD_180648_map_resolution2_x)
+				return;
+			v26 = -1;
+			a2 += (__int16)(a1 - x_DWORD_180648_map_resolution2_x) * (__int16)(a4 - a2) / (__int16)(a1 - a3);
+			a1 = x_DWORD_180648_map_resolution2_x - 1;
+			if ((a3 & 0x8000u) != 0)
+			{
+				a4 -= (__int16)-a3 * (__int16)(a4 - a2) / (__int16)(a1 - a3);
+				a3 = 0;
+			}
+		}
+	}
+	else
+	{
+		if ((a3 & 0x8000u) != 0)
+			return;
+		v26 = 1;
+		a2 += (__int16)-a1 * (__int16)(a4 - a2) / (__int16)(a3 - a1);
+		a1 = 0;
+		if ((__int16)(a3 - x_DWORD_180648_map_resolution2_x) >= 0)
+		{
+			a4 = a2 + (__int16)x_DWORD_180648_map_resolution2_x * (__int16)(a4 - a2) / (__int16)a3;
+			a3 = x_DWORD_180648_map_resolution2_x - 1;
+		}
+	}
+	v10 = (_BYTE*)(v27 + screenWidth_18062C * a2 + a1);
+	v11 = v26;
+	v12 = v26 * (a3 - a1);
+	v13 = a4 - a2;
+	if (a4 == a2)
+	{
+		v18 = a1;
+		v19 = a3;
+		if (a3 >= a1)
+		{
+			LOWORD(v19) = a3 - a1;
+		}
+		else
+		{
+			LOWORD(v18) = a3;
+			LOWORD(v19) = a1 - a3;
+		}
+	LABEL_54:
+		LOWORD(v19) = v19 + 1;
+		memset((void*)(v27 + screenWidth_18062C * a2 + v18), a5, v19);
+		return;
+	}
+	if (v13 <= v12)
+	{
+		v25 = screenWidth_18062C;
+	}
+	else
+	{
+		v12 = a4 - a2;
+		v13 = v26 * (a3 - a1);
+		v26 = screenWidth_18062C;
+		v25 = v11;
+	}
+	v14 = v12;
+	v15 = 2 * v13;
+	v16 = 2 * v13 - v12;
+	v17 = 2 * (v13 - v12);
+	*v10 = a5;
+	do
+	{
+		while (1)
+		{
+			v10 += v26;
+			if (v16 >= 0)
+				break;
+			v16 += v15;
+			*v10 = a5;
+			if (!--v14)
+				return;
+		}
+		v10 += v25;
+		v16 += v17;
+		*v10 = a5;
+		--v14;
+	} while (v14);
+}
